Compute D-optimal Sample Times

Computes D-optimal sampling times.

Usage

Dopt(run, data, clean = T)

Arguments

run

Run number of an NPAG model run that you wish to use to calculate the D-optimal sample times. An error will result if you try to use a run that was not NPAG.

data

An optional character vector with the filename of a Pmetrics data file which serves as the template for the dosage regimen and the initial times for the D-optimal sampling. The format of this file is the same as for any other Pmetrics run, except that EVID=0 observations are the initial sampling times. OUT values for these lines will be ignored. Only the first subject will be used as a template. If a Pmetrics data file is not specified, the first subject will be used from the original run as a template.

clean

Boolean operator to clean (delete) temporary files made during the optimization.

Value

A list of class PMdopt with 2 items.

allDopt

A data frame with 5 columns: timenum, time1, time2, gridpt, prob. The timenum column contains a number from 1 to the number of D-optimal sample times. The time1 and time2 columns contain the optimal times for each gridpoint and run. Two separate optimizations are made for stability checking so that the results can be compared. The gridpt column contains the gridpoint number for each set of optimal sample times. The prob column reports the probability of each gridpoint, and thus of each set of optimal sample times.

means

A data frame with the weighted mean optimal sample times for each of the two runs.

Details

This algorithm calculates the D-optimal sample times for each grid point in a nonparametric population model. It returns all the sample times and the mean weighted results. The optimization is run twice for stability checking, and the results of both runs are reported.

See also

Author

Michael Neely