Parse Pmetrics IT2B Output

ITparse processes the output from an IT2B run into a list.

Usage

ITparse(outfile = "IT_RF0001.TXT")

Arguments

outfile

This is the filename of the output from IT2B. Typically, the file will be called IT_RF0001.txt, and this is the default.

Value

The output of ITparse is a list with the following objects and of the class IT2B.

nsub

Number of subjects

nvar

Number of random variables or parameters in the model

nofix

Number of fixed variables or parameters in the model

par

Names of random parameters

parfix

Names of fixed parameters

covnames

Names of covariates

ab

Suggested boundaries for each random parameter to be passed to NPAG

fixedpos

Index of variables fixed to be positive

valfix

Values for fixed parameters

icycmax

Maximum number of cycles specified by the user

icyctot

Number of cycles run. If less than icycmax, convergence occurred.

stoptol

Stopping tolerance for convergence, default 0.001

converge

Boolean value if convergence occurred.

ODEtol

Ordindary Differential Equation solver tolerance.

numeqt

Number of output equations

ERRmod

Vector of length equal to numeqt whose values are 0 if gamma was estimated for that output equation or 1 if gamma was fixed to 1 for that output equation

ndrug

Number of drug inputs

salt

Vector of values of the salt fraction for each ndrug

ndose

Vector of the number of doses for each subject in the population

ncov

Number of covariates in the model

nobs

Vector of the number of observations for each subject in the population

nobsmax

Maximum number of observation in any individual subject

ypredpop

Array of population model predictions for each subject at each observation time point. ypredpopnsub,numeqt,time,type where type is 1=mean, 2=median of the population prior used to calculate ypredpop

ypredbay

Array of Bayesian posterior model predictions for each subject at each observation time point. ypredbaynsub,numeqt,time,type where type is 1=mean, 2=median of the population prior used to calculate ypredbay

parbay

Array of Bayesian posterior parameter estimates for each subject, parbaynsub,nvar,type where type is 1=mean, 2=median of the population prior used to calculate parbay

ic

Data frame with one row and two columns for final cycle Akaike Information Criterion (AIC) and Bayesian Information Criterion (BIC)

ilog

Vector of cycle number and associated log-likelihood

imean

Matrix of cycle numbers and associated means for each random parameter

imed

Matrix of cycle numbers and associated medians for each random parameter

isd

Matrix of cycle numbers and associated standard deviations for each random parameter

icv

Matrix of cycle numbers and associated coefficients of variation for each random parameter

igamlam

Matrix of cycle number and associated gamma or lambda with each output equation in a column

lpar

Matrix of subjects in rows and MAP Bayesian parameter estimates in columns for each parameter, based on population means from the next to last cycle.

lsd

Matrix of subjects in rows and SD of Bayesian posterior parameter distributions in columns for each parameter, based on population means from the next to last cycle.

lcv

Matrix of subjects in rows and CV of Bayesian posterior parameter distributions in columns for each parameter, based on population means from the next to last cycle.

sdata

Subject data consisting of 5 columns: id, nsub, age, sex, ht, id is the original identification number in the .csv matrix file; nsub is the sequential subject number in the IT2B run; age, sex and ht will be missing for .csv input and present if included in .wrk input files

dosecov

Data frame with all dosing information for each subject, including times, routes, amounts, and associated covariate values

outputs

Data frame with measured outputs for each subject and associated assay error polynomials. The order of the columns is nsub, time, numeqt, observation, c0, c1, c2, c3, where the last four columns are the coefficients of the assay error polynomial for that observation, such that SDobs = c0 + c1*obs + c2*obs*2 + c3obs**3

negflag

A flag indicating that some negative predictions were changed to missing. This means that the model may be misspecified.

mdata

The filename of the data used in the run.

Details

This function can take some time to process the RFILE, depending on the number of subjects, doses, observations, etc. Typical wait times are a few seconds up to 5 minutes. When processing is complete a summary of the extracted data will be reported on the console.

Author

Michael Neely