Compare NPAG or IT2B runs

r lifecycle::badge("stable")

Compare parameters, convergence, -2*log likelihood, AIC and bias and imprecision of population and posterior predictions.

Usage

PM_compare(x, y, ..., icen = "median", outeq = 1, plot = TRUE)

Arguments

x

The first PM_result object you wish to compare, already loaded with PM_load. This will serve as the reference output for P-value testing (see details).

y

The second PM_result object to compare.

...

Additional PM_result objects to compare. See details. Also, parameters to be passed to plot.PM_op if plot is true as well as to mtsknn.eq. Order does not matter.

icen

Can be either "median" for the predictions based on medians of pred.type parameter value distributions, or "mean". Default is "median".

outeq

Number of the output equation to compare; default is 1.

plot

Boolean operator selecting whether to generate the report.

Value

If plot = TRUE, a report will be generated comparing the selected models. Additionally, a data frame with the following objects for each model is invisbily returned.

run The run number of the data

nsub Number of subjects in the model

nvar Number of random parameters in the model

par Names of random parameters

cycles Number of cycles run

converge Boolean value if convergence occurred.

ll Final cycle -2*Log-likelihood

One of the following, depending on the option set in setPMoptions:

aic Final cycle Akaike Information Criterion OR

bic Final cycle Bayesian (Schwartz) Information Criterion

popBias Bias, calculated by the method set in setPMoptions, of the predictions based on icen population parameters

popImp Imprecision, calculated by the method set in setPMoptions, of the predictions based on icen population parameters

postBias Bias, calculated by the method set in setPMoptions, of the predictions based on icen posterior parameters

postImp Imprecision, calculated by the method set in setPMoptions, of the predictions based on icen posterior parameters

pval P-value for each model compared to the first. See details.

Details

Objects can be specified separated by commas, e.g. PM_compare(run1, run2, run3) followed by any arguments you wish to plot.PM_op, mtsknn.eq. P-values are based on comparison using the nearest neighbors approach if all models are non-parametrics. Models may only be compared on parameters that are included in the first model. The P-value is the comparison between each model and the first model in the list. Missing P-values are when a model has no parameter names in common with the first model, and for the first model compared to itself. Significant P-values indicate that the null hypothesis should be rejected, i.e. the joint distributions between the two compared models for that parameter are significantly different.

See also

Author

Michael Neely