Contains covariate data

Contains a data frame with subject-specific covariate data output.

Details

The PM_cov object is both a data field within a PM_result, and itself an R6 object comprising data fields and associated methods suitable for analysis and plotting relationships between covariates and posterior parameters, covariates over time, or parameter values over time.

Because PM_cov objects are automatically added to the PM_result at the end of a successful run, it is generally not necessary for users to generate PM_cov objects themselves.

The results are contained in the $data field, and it is this field which is passed to the $plot and $summary methods. You can use this $data field for custom manipulations, although usually this is better done on the summary object, e.g., run1$cov$summary() %>% select(africa, gender) %>% table(). If you are unfamiliar with the %>% pipe function, please type help("%>%", "magrittr") into the R console and look online for instructions/tutorials in tidyverse, a powerful approach to data manipulation upon which Pmetrics is built.

This output of this function is suitable for exploration of covariate- parameter, covariate-time, or parameter-time relationships.

See also

Author

Michael Neely, Julian Otalvaro

Public fields

data

A data frame with the following columns

id Subject identification

time Times of covariate observations

covnames... Columns with each covariate observations in the dataset for each subject and time

parnames... Columns with each parameter in the model and the icen summary for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian parameter means and medians

icen The type of summarized Bayesian posterior individual parameter values: mean or median

Methods

Public methods

PM_cov$new()

PM_cov$step()

PM_cov$summary()

PM_cov$plot()

PM_cov$print()

PM_cov$clone()

Method new()

Create new object populated with covariate-parameter information

Usage

PM_cov$new(PMdata = NULL, path = ".", ...)

Arguments

PMdata

include Saved, parsed output of prior run, used when source files are not available. .

path

include Path to the folder containing the raw results of the run. Default is the current working directory. .

...

Not currently used.

Details

Creation of new PM_cov object is automatic and not generally necessary for the user to do.

Method step()

Stepwise linear regression of covariates and Bayesian posterior parameter values

Usage

PM_cov$step(...)

Arguments

...

Arguments passed to PM_step

Details

See PM_step.

Method summary()

Summary method

Usage

PM_cov$summary(...)

Arguments

...

Arguments passed to summary.PM_cov

Details

See summary.PM_cov.

Method plot()

Plot method

Usage

PM_cov$plot(...)

Arguments

...

Arguments passed to plot.PM_cov

Details

See plot.PM_cov.

Method print()

Print method

Usage

PM_cov$print(...)

Arguments

...

Arguments passed to print

Details

Print method for PM_cov

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_cov$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.