Pmetrics Run Cycle Information

Pmetrics Run Cycle Information

Pmetrics Run Cycle Information

Details

Contains the cycle information after a run.

This contains the output of makeCycle after a run, which generates information suitable for analysis and plotting of cycle information. Each field corresponds to a column in the complete data frame.

To manipulate the entire data frame, use the data field, e.g. final <- run1$cycle$data %>% slice_tail(n=1). If you are unfamiliar with the %>% pipe function, please type help("%>%", "magrittr") into the R console and look online for instructions/tutorials in tidyverse, a powerful approach to data manipulation upon which Pmetrics is built.

Creation of new PM_cycle object is automatic and not generally necessary for the user to do.

See plot.PMcycle.

Public fields

names

Vector of names of the random parameters

cynum

Vector cycle numbers, which may start at numbers greater than 1 if a non-uniform prior was specified for the run (NPAG only)

ll

Matrix of cycle number and -2*Log-likelihood at each cycle

gamlam

A matrix of cycle number and gamma or lambda at each cycle

mean

A matrix of cycle number and the mean of each random parameter at each cycle, normalized to initial mean

sd

A matrix of cycle number and the standard deviation of each random parameter at each cycle, normalized to initial standard deviation

median

A matrix of cycle number and the median of each random parameter at each cycle, normalized to initial median

aic

A matrix of cycle number and Akaike Information Criterion at each cycle

bic

A matrix of cycle number and Bayesian (Schwartz) Information Criterion at each cycle

data

A data frame combining all the above fields as its columns

Methods

Public methods

PM_cycle$new()

PM_cycle$plot()

PM_cycle$clone()

Method new()

Create new object populated with cycle information

Usage

PM_cycle$new(cycle)

Arguments

cycle

The parsed output from makeCycle.

Method plot()

Plot method

Usage

PM_cycle$plot(...)

Arguments

...

Arguments passed to plot.PMcycle

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_cycle$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.