Add Jelliffe Creatinine Clearance

Gets creatinine clearance as estimated by the Jelliffe equation.

Usage

PMgetCRCL(
  mdata,
  idCol = "id",
  wtCol = "wt",
  maleCol = "male",
  ageCol = "age",
  scrCol = "scr",
  SI = F
)

Arguments

mdata

A Pmetrics matrix data object

idCol

A character vector with the name of the id column in mdata. The default is “id”.

wtCol

A character vector with the name of the weight column in data. The default is “wt”.

maleCol

A character vector with the name of the gender column in mdata. Male should be 1 and female should be 0. The default is “male”.

ageCol

A character vector with the name of the age column in mdata. The default is “age”.

scrCol

A character vector with the name of the serum creatinine column in mdata. Default units are mg/dL, and the the default name is “scr”.

SI

Boolean value, if true, will expect serum creatinine to be in micromol/L. Default is FALSE.

Value

A vector of length nrow(mdata) with Jelliffe CRCL values for every dose in mdata. Vector values for observation events in mdata are NA.

Details

The equation depends on age, sex, weight, and serum creatinine. ESS = wt * (29.3 - (0.203 * age)) for males ESS = wt * (25.1 - (0.175 * age)) for females scrAve = (Scr1 + Scr2) / 2 ESS_cor = ESS * (1.035 - (0.0337 * scrAve)) E = ESS_cor - 4 * wt * (Scr2 - Scr1) / (time2 - time1) CRCL = E / (14.4 * scrAve) in ml/min/1.73m^2

Author

Michael Neely