Summarize NPAG or IT2B Run

Generates a summary of a Pmetrics NPAG or IT2B run

Usage

PMreport(wd, rdata, icen = "median", type = "NPAG", parallel = F)

Arguments

wd

The working directory containing the NP_RFxxxx.TXT or IT_RFxxxx.TXT file

rdata

The processed output of an IT2B or NPAG run, depending on local or server runs.

icen

Median (default), mean or mode of Bayesian posterior to be used to calculate predictions.

type

“NPAG” (default) or “IT2B” report type

parallel

Boolean parameter which indicates the type of run done. Default is FALSE for serial.

Value

Several files are placed in the wd

NPAGreport.html or IT2Breport.html

An .html file containing a summary of all the results

poppoints.csv

NPAG only: A .csv file containing the population support points and probabilities

poparam.csv

A .csv file containing a summary of the population parameter values, including mean, standard deviation, coefficient of variation, variance, and median

popcor.csv

A .csv file containing the population parameter correlation matrix

popcov.csv

A .csv file containing the population parameter covariance matrix

cycle.pdf

A .pdf file containing the run cycle information (see plot.PMcycle)

cycle.png

A thumbnail of the run cycle information for the .html file

final.pdf

A .pdf file containing the population final cycle information (see plot.PMfinal)

final.png

A thumbnail of the population final cycle information for the .html file

opx.pdf

One or more .pdf files, where x is the number of the output equation, each containing two observed vs. predicted plots: population and individual Bayesian posterior predictions (see plot.PMop)

opx.png

One or more thumnails of the observed vs. predicted plots for the .html file

NPAGout.Rdata or IT2Bout.Rdata

An R data file containing the output of NPparse or ITparse, makeFinal, makeCycle, makeOP, makeCov, makePop, makePost, and the data file for the run read by PMreadMatrix. This file can be loaded using PMload.

Details

Creates an HTML page and several files summarizing an NPAG or IT2B run. This report is generated automatically at the end of a successful run.

Author

Michael Neely