Extract covariate information

Extracts covariate information from a Pmetrics data object.

Usage

getCov(mdata)

Arguments

mdata

A PM_data object. It can be other data frames but the results will likely not be meaningful.

Value

A list with named items: ncov, covnames, covstart, covend.

Details

The function subtracts the number of fixed columns in a standard data format, currently 14, from the total number of columns in mdata and queries the remaining columns. When given a PM_data object, will return a list with the number of covariates, their names, and the starting and ending column numbers

Author

Michael Neely

Examples

if (FALSE) { # \dontrun{
getCov(dataEx)
} # }