Extract covariate data

Generates an data.frame with subject-specific covariate data from an NPAG or IT2B object

Usage

makeCov(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeCov is a dataframe of class PMcov, which has the following columns:

id

Subject identification

time

Times of covariate observations

covnames...

Columns with each covariate observations in the dataset for each subject and time

parnames...

Columns with each parameter in the model and the icen summary for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian parameter means and medians

icen

The type of summarized Bayesian posterior individual parameter values: mean or median.

Details

For each subject, makeCov extracts covariate information and Bayesian posterior parameter estimates. This output of this function is suitable for exploration of covariate-parameter, covariate-time, or parameter-time relationships.

See also

Author

Michael Neely

Examples

library(PmetricsData)
cov <- makeCov(NPex$NPdata)
cov
#>    id time   wt africa age gender height block       Ka        Ke        V
#> 1   1    0 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 2   1   24 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 3   1   48 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 4   1   72 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 5   1   96 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 6   1  120 46.7      1  21      1    160     1 0.440192 0.0246167  66.3882
#> 7   2    0 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 8   2   24 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 9   2   48 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 10  2   72 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 11  2   96 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 12  2  120 66.5      1  30      1    174     1 0.738613 0.0398178 119.4690
#> 13  3    0 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 14  3   24 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 15  3   48 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 16  3   72 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 17  3   96 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 18  3  120 46.7      1  24      0    164     1 0.899944 0.0431027 108.6490
#> 19  4    0 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 20  4   24 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 21  4   48 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 22  4   72 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 23  4   96 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 24  4  120 50.8      1  25      1    165     1 0.897550 0.0564307 119.8190
#> 25  5    0 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 26  5   24 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 27  5   48 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 28  5   72 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 29  5   96 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 30  5  120 65.8      1  22      1    181     1 0.105273 0.0675058 113.3430
#> 31  6    0 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 32  6   24 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 33  6   48 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 34  6   72 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 35  6   96 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 36  6  120 65.0      1  23      1    177     1 0.895237 0.0348831  71.8617
#> 37  7    0 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 38  7   24 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 39  7   48 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 40  7   72 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 41  7   96 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 42  7  120 51.7      1  27      0    161     1 0.215198 0.0832836  35.2243
#> 43  8    0 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 44  8   24 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 45  8   48 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 46  8   72 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 47  8   96 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 48  8  120 51.2      1  22      1    163     1 0.895494 0.0348882  71.8466
#> 49  9    0 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 50  9   24 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 51  9   48 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 52  9   72 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 53  9   96 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 54  9  120 55.0      1  23      1    174     1 0.790042 0.0439417 101.7520
#> 55 10    0 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 56 10   24 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 57 10   48 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 58 10   72 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 59 10   96 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 60 10  120 52.1      1  32      1    163     1 0.655770 0.0615898  61.6882
#> 61 11    0 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 62 11   24 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 63 11   48 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 64 11   72 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 65 11   96 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 66 11  120 56.5      1  34      1    165     1 0.583214 0.0683228  73.1084
#> 67 12    0 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 68 12   24 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 69 12   48 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 70 12   72 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 71 12   96 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 72 12  120 47.9      1  54      0    160     1 0.470094 0.0306855  91.8654
#> 73 13    0 60.5      1  24      1    180     1 0.215198 0.0832837  35.2243
#> 74 13   24 60.5      1  24      1    180     1 0.215198 0.0832837  35.2243
#> 75 13   48 60.5      1  24      1    180     1 0.215198 0.0832837  35.2243
#> 76 13   72 60.5      1  24      1    180     1 0.215198 0.0832837  35.2243
#>          lag icen
#> 1  0.5544030 mean
#> 2  0.5544030 mean
#> 3  0.5544030 mean
#> 4  0.5544030 mean
#> 5  0.5544030 mean
#> 6  0.5544030 mean
#> 7  0.0259495 mean
#> 8  0.0259495 mean
#> 9  0.0259495 mean
#> 10 0.0259495 mean
#> 11 0.0259495 mean
#> 12 0.0259495 mean
#> 13 2.0959200 mean
#> 14 2.0959200 mean
#> 15 2.0959200 mean
#> 16 2.0959200 mean
#> 17 2.0959200 mean
#> 18 2.0959200 mean
#> 19 0.6883040 mean
#> 20 0.6883040 mean
#> 21 0.6883040 mean
#> 22 0.6883040 mean
#> 23 0.6883040 mean
#> 24 0.6883040 mean
#> 25 0.0186591 mean
#> 26 0.0186591 mean
#> 27 0.0186591 mean
#> 28 0.0186591 mean
#> 29 0.0186591 mean
#> 30 0.0186591 mean
#> 31 1.9978800 mean
#> 32 1.9978800 mean
#> 33 1.9978800 mean
#> 34 1.9978800 mean
#> 35 1.9978800 mean
#> 36 1.9978800 mean
#> 37 1.7966900 mean
#> 38 1.7966900 mean
#> 39 1.7966900 mean
#> 40 1.7966900 mean
#> 41 1.7966900 mean
#> 42 1.7966900 mean
#> 43 1.9985200 mean
#> 44 1.9985200 mean
#> 45 1.9985200 mean
#> 46 1.9985200 mean
#> 47 1.9985200 mean
#> 48 1.9985200 mean
#> 49 0.8801720 mean
#> 50 0.8801720 mean
#> 51 0.8801720 mean
#> 52 0.8801720 mean
#> 53 0.8801720 mean
#> 54 0.8801720 mean
#> 55 0.8013670 mean
#> 56 0.8013670 mean
#> 57 0.8013670 mean
#> 58 0.8013670 mean
#> 59 0.8013670 mean
#> 60 0.8013670 mean
#> 61 1.3385500 mean
#> 62 1.3385500 mean
#> 63 1.3385500 mean
#> 64 1.3385500 mean
#> 65 1.3385500 mean
#> 66 1.3385500 mean
#> 67 1.0252500 mean
#> 68 1.0252500 mean
#> 69 1.0252500 mean
#> 70 1.0252500 mean
#> 71 1.0252500 mean
#> 72 1.0252500 mean
#> 73 1.7957700 mean
#> 74 1.7957700 mean
#> 75 1.7957700 mean
#> 76 1.7957700 mean
#>  [ reached 'max' / getOption("max.print") -- omitted 164 rows ]
names(cov)
#>  [1] "id"     "time"   "wt"     "africa" "age"    "gender" "height" "block" 
#>  [9] "Ka"     "Ke"     "V"      "lag"    "icen"