Summarize NPAG or IT2B Final Cycle Population Values

Extracts final cycle information from NPAG or IT2B run.

Usage

makeFinal(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeFinal is a list of class PMfinal, which contains the following:

popPoints

(NPAG only) Dataframe of the final cycle joint population density of grid points with column names equal to the name of each random parameter plus prob for the associated probability of that point

popMean

The final cycle mean for each random parameter distribution

popSD

The final cycle standard deviation for each random parameter distribution

popCV

The final cycle coefficient of variation (SD/Mean) for each random parameter distribution

popVar

The final cycle variance for each random parameter distribution

popCov

The final cycle random parameter covariance matrix

popCor

The final cycle random parameter correlation matrix

popMedian

The final cycle median values for each random parameter

postPoints

(NPAG only) Dataframe of posterior population points for each of the first 100 subject, with columns id, point, parameters and probability. The first column is the subject, the second column has the population point number, followed by the values for the parameters in that point and the probability.

postMean

A nsub x npar data frame containing the means of the posterior distributions for each parameter.

postSD

A nsub x npar data frame containing the SDs of the posterior distributions for each parameter.

postVar

A nsub x npar data frame containing the variances of the posterior distributions for each parameter.

postCov

NPAG only: An array of dimensions npar x npar x nsub that contains the covariances of the posterior distributions for each parameter and subject.

postCor

NPAG only: An array of dimensions npar x npar x nsub that contains the correlations of the posterior distributions for each parameter and subject.

postMed

A nsub x npar data frame containing the medians of the posterior distributions for each parameter.

shrinkage

A data frame with the shrinkage for each parameter. The total population variance for a parameter is comprised of variance(EBE) plus average variance(EBD), where each subject's EBE is the Empirical Bayes Estimate or mean posterior value for the parameter. EBD is the Empirical Bayes Distribution, or the full Bayesian posterior parameter value distribution for each subject.

The typical definition of \(\eta\) shrinkage is \([1 - \frac{SD(\eta)}{\omega}]\) or \([1 - \frac{var(\eta)}{\omega^2}]\), where \(\eta\) is the EBE and \(\omega^2\) is the population variance of \(\eta\).

In parametric modeling approaches \(\eta\) is the interindividual variability around the typical (mean) value of the parameter in the population, usually referred to as \(\theta\). In nonparametric approaches, there is no assumption of normality, so \(\eta\) simply becomes each subject's mean parameter value estimate.

Here is how Pmetrics derives and then calculates shrinkage for a given parameter. $$popVar = var(EBE) + mean(var(EBD))$$ $$1 = \frac{var(EBE)}{popVar} + \frac{mean(var(EBD)}{popVar}$$ $$1 - \frac{var(EBE)}{popVar} = \frac{mean(var(EBD))}{popVar}$$ $$shrinkage = \frac{mean(var(EBD))}{popVar}$$ Shrinkage is therefore a fraction between 0 and 1. If Bayesian posterior distributions are wide for a given parameter and \(mean(var(EBD))\) is high due to sparse or uninformative sampling, then most of the population variance is due to this variance and shrinkage is high, i.e., individual posterior estimates (EBE) shrink towards the population mean. Be aware, however, that a Bayesian posterior parameter value distribution for a given subject who is sparsely sampled may also be a single support point with no variance. Therefore EBD under nonparametric assumptions is not always large with uninformative sampling. This means that shrinkage is not as readily interpretable in nonparametric population modeling.

An alternative is to consider the number of support points relative to the number of subjects. Highly informed, distinct subjects will result in the maximum possible number of support points, N, which is the same as the number of subjects. In contrast, badly undersampled subjects can result in only one support point. There is no formal criterion for this statistic, but it can be used in combination with shrinkage to assess the information content of the data.

gridpts

(NPAG only) Initial number of support points

nsub

Number of subjects

ab

Matrix of boundaries for random parameter values

A plot method exists in plot for PMfinal objects.

Details

This function will parse the output of NPparse or ITparse to generate a list suitable for analysis and plotting of NPAG or IT2B final cycle population values.

See also

Author

Michael Neely

Examples

final <- makeFinal(NPex$NPdata)
final
#> $popPoints
#>           Ka         Ke         V      Tlag1       prob
#> 1  0.8999440 0.04310272 108.64920 2.09592004 0.05000771
#> 2  0.2151984 0.08328365  35.22432 1.80379203 0.10419106
#> 3  0.7927824 0.04393907 101.57052 0.88691202 0.04985577
#> 4  0.6552048 0.06166284  61.52304 0.80102401 0.04980239
#> 5  0.6575952 0.06211706  30.98496 1.92417603 0.04999970
#> 6  0.8952615 0.02318994  75.94093 1.76207524 0.05076679
#> 7  0.2151984 0.08328365  35.22432 1.77879203 0.04581144
#> 8  0.4324662 0.02464088  66.22597 0.53389922 0.04923321
#> 9  0.8976352 0.05642951 119.81946 0.68837602 0.05021143
#> 10 0.7540983 0.03531704  89.66107 0.68665921 0.05077266
#> 11 0.8008512 0.03427034  71.83326 1.04925602 0.04863507
#> 12 0.7477824 0.03960782 119.57052 0.01191201 0.04798899
#> 13 0.5527824 0.04393907 117.88302 0.28691201 0.05234436
#> 14 0.5829040 0.06831802  73.11720 1.33872003 0.05074312
#> 15 0.8963072 0.07349612  63.15906 1.17733602 0.04925686
#> 16 0.4608512 0.03055784  92.08326 1.02425602 0.04980933
#> 17 0.8958512 0.03488909  71.83326 1.99925603 0.10078061
#> 18 0.1018374 0.06755382 113.31523 0.01575520 0.04978951
#> 
#> $popMean
#>          Ka         Ke       V    Tlag1
#> 1 0.6282269 0.05138859 77.7699 1.185559
#> 
#> $popSD
#>        Ka         Ke        V     Tlag1
#> 1 0.26297 0.01960317 28.82353 0.6677845
#> 
#> $popCV
#>         Ka       Ke        V    Tlag1
#> 1 41.85908 38.14694 37.06258 56.32656
#> 
#> $popVar
#>           Ka           Ke        V     Tlag1
#> 1 0.06915324 0.0003842845 830.7959 0.4459361
#> 
#> $popCov
#>                 Ka            Ke           V         Tlag1
#> Ka     0.069153244 -0.0029031355   2.5298201   0.026000528
#> Ke    -0.002903136  0.0003842845  -0.2704453   0.002291387
#> V      2.529820102 -0.2704453298 830.7959212 -12.200365241
#> Tlag1  0.026000528  0.0022913867 -12.2003652   0.445936089
#> 
#> $popCor
#>               Ka         Ke          V      Tlag1
#> Ka     1.0000000 -0.5631637  0.3337615  0.1480606
#> Ke    -0.5631637  1.0000000 -0.4786366  0.1750393
#> V      0.3337615 -0.4786366  1.0000000 -0.6338542
#> Tlag1  0.1480606  0.1750393 -0.6338542  1.0000000
#> 
#> $popMedian
#>          Ka         Ke        V    Tlag1
#> 2 0.7026888 0.04393907 72.47523 1.113296
#> 
#> $popRanFix
#> NULL
#> 
#> $postPoints
#> # A tibble: 96 × 7
#>       id point    Ka     Ke     V  Tlag1     prob
#>    <dbl> <int> <dbl>  <dbl> <dbl>  <dbl>    <dbl>
#>  1     1     1 0.895 0.0232  75.9 1.76   1.67e- 2
#>  2     1     2 0.432 0.0246  66.2 0.534  9.83e- 1
#>  3     2     1 0.900 0.0431 109.  2.10   7.75e- 8
#>  4     2     2 0.793 0.0439 102.  0.887  1.18e- 3
#>  5     2     3 0.748 0.0396 120.  0.0119 9.52e- 1
#>  6     2     4 0.553 0.0439 118.  0.287  4.73e- 2
#>  7     2     5 0.583 0.0683  73.1 1.34   2.13e-10
#>  8     3     1 0.900 0.0431 109.  2.10   1.00e+ 0
#>  9     3     2 0.583 0.0683  73.1 1.34   7.70e- 9
#> 10     4     1 0.898 0.0564 120.  0.688  1.00e+ 0
#> # ℹ 86 more rows
#> 
#> $postMean
#>    id       Ka        Ke        V     Tlag1
#> 1   1 0.440192 0.0246167  66.3882 0.5544030
#> 2   2 0.738613 0.0398178 119.4690 0.0259495
#> 3   3 0.899944 0.0431027 108.6490 2.0959200
#> 4   4 0.897550 0.0564307 119.8190 0.6883040
#> 5   5 0.105273 0.0675058 113.3430 0.0186591
#> 6   6 0.895237 0.0348831  71.8617 1.9978800
#> 7   7 0.215198 0.0832836  35.2243 1.7966900
#> 8   8 0.895494 0.0348882  71.8466 1.9985200
#> 9   9 0.790042 0.0439417 101.7520 0.8801720
#> 10 10 0.655770 0.0615898  61.6882 0.8013670
#> 11 11 0.583214 0.0683228  73.1084 1.3385500
#> 12 12 0.470094 0.0306855  91.8654 1.0252500
#> 13 13 0.215198 0.0832837  35.2243 1.7957700
#> 14 14 0.555242 0.0439039 117.8380 0.2868650
#> 15 15 0.795724 0.0342102  72.1966 1.0584400
#> 16 16 0.752981 0.0352990  89.6702 0.6879460
#> 17 17 0.891343 0.0734141  63.3168 1.1798900
#> 18 18 0.894633 0.0231919  75.9277 1.7604100
#> 19 19 0.657597 0.0621169  30.9852 1.9241800
#> 20 20 0.215198 0.0832837  35.2243 1.7960100
#> 
#> $postSD
#>    id          Ka          Ke          V       Tlag1
#> 1   1 5.92949e-02 1.85900e-04 1.24471000 1.57358e-01
#> 2   2 4.14333e-02 9.30204e-04 0.71152100 6.54383e-02
#> 3   3 2.78188e-05 2.21253e-06 0.00311777 6.64408e-05
#> 4   4 8.25437e-03 1.15386e-04 0.06746470 6.97672e-03
#> 5   5 5.21762e-02 7.29362e-04 0.42644700 4.41001e-02
#> 6   6 1.63377e-02 1.72654e-04 0.76091100 3.65945e-02
#> 7   7 2.13211e-04 1.51594e-05 0.01146760 1.12733e-02
#> 8   8 1.27877e-02 3.57297e-04 0.60893900 2.66606e-02
#> 9   9 2.55413e-02 2.46894e-04 1.81722000 6.57720e-02
#> 10 10 8.80413e-03 1.13586e-03 2.56904000 6.12762e-03
#> 11 11 9.85709e-03 1.69250e-04 0.36892400 5.74677e-03
#> 12 12 5.57671e-02 7.57640e-04 1.87978000 9.73359e-02
#> 13 13 0.00000e+00 0.00000e+00 0.00000000 0.00000e+00
#> 14 14 2.26544e-02 3.94219e-04 0.99697500 4.43318e-02
#> 15 15 4.62149e-02 4.97149e-04 2.68764000 9.53561e-02
#> 16 16 1.80654e-02 2.97213e-04 0.15839500 2.08028e-02
#> 17 17 3.91285e-02 6.46488e-04 1.24328000 2.01489e-02
#> 18 18 1.70457e-02 5.42301e-05 0.35797400 4.52360e-02
#> 19 19 5.89857e-04 6.74131e-05 0.10113700 1.88146e-04
#> 20 20 4.54947e-05 2.66963e-06 0.00487837 1.15747e-02
#> 
#> $postVar
#>    id           Ka           Ke            V        Tlag1
#> 1   1 3.515885e-03 3.455881e-08 1.549303e+00 2.476154e-02
#> 2   2 1.716718e-03 8.652795e-07 5.062621e-01 4.282171e-03
#> 3   3 7.738856e-10 4.895289e-12 9.720490e-06 4.414380e-09
#> 4   4 6.813462e-05 1.331393e-08 4.551486e-03 4.867462e-05
#> 5   5 2.722356e-03 5.319689e-07 1.818570e-01 1.944819e-03
#> 6   6 2.669204e-04 2.980940e-08 5.789855e-01 1.339157e-03
#> 7   7 4.545893e-08 2.298074e-10 1.315058e-04 1.270873e-04
#> 8   8 1.635253e-04 1.276611e-07 3.708067e-01 7.107876e-04
#> 9   9 6.523580e-04 6.095665e-08 3.302289e+00 4.325956e-03
#> 10 10 7.751271e-05 1.290178e-06 6.599967e+00 3.754773e-05
#> 11 11 9.716222e-05 2.864556e-08 1.361049e-01 3.302537e-05
#> 12 12 3.109969e-03 5.740184e-07 3.533573e+00 9.474277e-03
#> 13 13 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 14 14 5.132218e-04 1.554086e-07 9.939592e-01 1.965308e-03
#> 15 15 2.135817e-03 2.471571e-07 7.223409e+00 9.092786e-03
#> 16 16 3.263587e-04 8.833557e-08 2.508898e-02 4.327565e-04
#> 17 17 1.531040e-03 4.179467e-07 1.545745e+00 4.059782e-04
#> 18 18 2.905559e-04 2.940904e-09 1.281454e-01 2.046296e-03
#> 19 19 3.479313e-07 4.544526e-09 1.022869e-02 3.539892e-08
#> 20 20 2.069768e-09 7.126924e-12 2.379849e-05 1.339737e-04
#> 
#> $postCov
#> , , subj = 1
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     3.515885e-03 -1.102292e-05  0.0738051877  9.330533e-03
#>   Ke    -1.102292e-05  3.455878e-08 -0.0002313922 -2.925286e-05
#>   V      7.380519e-02 -2.313922e-04  1.5493126444  1.958658e-01
#>   Tlag1  9.330533e-03 -2.925286e-05  0.1958658182  2.476157e-02
#> 
#> , , subj = 2
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     1.716722e-03 -3.779069e-05  0.0136823127 -2.361088e-03
#>   Ke    -3.779069e-05  8.652787e-07 -0.0004163413  5.785253e-05
#>   V      1.368231e-02 -4.163413e-04  0.5062625240 -3.909039e-02
#>   Tlag1 -2.361088e-03  5.785253e-05 -0.0390903889  4.282170e-03
#> 
#> , , subj = 3
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     7.738863e-10 -6.154988e-11  8.673268e-08  1.848305e-09
#>   Ke    -6.154988e-11  4.895277e-12 -6.898153e-09 -1.470022e-10
#>   V      8.673268e-08 -6.898153e-09  9.720494e-06  2.071473e-07
#>   Tlag1  1.848305e-09 -1.470022e-10  2.071473e-07  4.414385e-09
#> 
#> , , subj = 4
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     6.813454e-05 -9.524405e-07  5.568784e-04  5.758839e-05
#>   Ke    -9.524405e-07  1.331399e-08 -7.784503e-06 -8.050177e-07
#>   V      5.568784e-04 -7.784503e-06  4.551489e-03  4.706824e-04
#>   Tlag1  5.758839e-05 -8.050177e-07  4.706824e-04  4.867461e-05
#> 
#> , , subj = 5
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     2.722356e-03 -3.805533e-05  0.0222504089  2.300978e-03
#>   Ke    -3.805533e-05  5.319685e-07 -0.0003110345 -3.216496e-05
#>   V      2.225041e-02 -3.110345e-04  0.1818574305  1.880640e-02
#>   Tlag1  2.300978e-03 -3.216496e-05  0.0188063963  1.944823e-03
#> 
#> , , subj = 6
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     2.669195e-04  2.602925e-06 -0.0123577534  5.970861e-04
#>   Ke     2.602925e-06  2.980950e-08 -0.0001259876  5.927812e-06
#>   V     -1.235775e-02 -1.259876e-04  0.5789859114 -2.777368e-02
#>   Tlag1  5.970861e-04  5.927812e-06 -0.0277736765  1.339157e-03
#> 
#> , , subj = 7
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     4.545896e-08 -3.232142e-09  2.445012e-06  1.352866e-08
#>   Ke    -3.232142e-09  2.298060e-10 -1.738409e-07 -9.618909e-10
#>   V      2.445012e-06 -1.738409e-07  1.315051e-04  7.276397e-07
#>   Tlag1  1.352866e-08 -9.618909e-10  7.276397e-07  1.270878e-04
#> 
#> , , subj = 8
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     1.635259e-04 -1.934125e-07 -0.0053576255  3.228863e-04
#>   Ke    -1.934125e-07  1.276614e-07 -0.0001514058  2.671457e-06
#>   V     -5.357626e-03 -1.514058e-04  0.3708063126 -1.435922e-02
#>   Tlag1  3.228863e-04  2.671457e-06 -0.0143592201  7.107889e-04
#> 
#> , , subj = 9
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     6.523574e-04 -4.142693e-07 -0.0430318592  1.641200e-03
#>   Ke    -4.142693e-07  6.095657e-08 -0.0001340919  9.836287e-08
#>   V     -4.303186e-02 -1.340919e-04  3.3022759234 -1.098566e-01
#>   Tlag1  1.641200e-03  9.836287e-08 -0.1098566296  4.325961e-03
#> 
#> , , subj = 10
#> 
#>        par2
#> par1               Ka            Ke            V         Tlag1
#>   Ka     7.751278e-05 -9.985336e-06  0.022568488  4.768098e-05
#>   Ke    -9.985336e-06  1.290186e-06 -0.002917729 -6.014794e-06
#>   V      2.256849e-02 -2.917729e-03  6.599960039  1.350484e-02
#>   Tlag1  4.768098e-05 -6.014794e-06  0.013504838  3.754774e-05
#> 
#> , , subj = 11
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     9.716231e-05  1.506918e-06 -2.761677e-03 -5.281318e-05
#>   Ke     1.506918e-06  2.864566e-08 -6.023641e-05 -6.712859e-07
#>   V     -2.761677e-03 -6.023641e-05  1.361046e-01  1.008036e-03
#>   Tlag1 -5.281318e-05 -6.712859e-07  1.008036e-03  3.302534e-05
#> 
#> , , subj = 12
#> 
#>        par2
#> par1               Ka            Ke            V         Tlag1
#>   Ka     3.109966e-03  4.109677e-05 -0.084613908  1.343227e-03
#>   Ke     4.109677e-05  5.740184e-07 -0.000925752  1.405407e-06
#>   V     -8.461391e-02 -9.257520e-04  3.533566089 -1.357273e-01
#>   Tlag1  1.343227e-03  1.405407e-06 -0.135727321  9.474270e-03
#> 
#> , , subj = 13
#> 
#>        par2
#> par1    Ka Ke V        Tlag1
#>   Ka     0  0 0 0.0000000000
#>   Ke     0  0 0 0.0000000000
#>   V      0  0 0 0.0000000000
#>   Tlag1  0  0 0 0.0001362316
#> 
#> , , subj = 14
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     5.132202e-04 -6.792630e-06 -1.137316e-02  9.105388e-05
#>   Ke    -6.792630e-06  1.554086e-07 -6.773826e-05  9.844952e-06
#>   V     -1.137316e-02 -6.773826e-05  9.939589e-01 -3.958429e-02
#>   Tlag1  9.105388e-05  9.844952e-06 -3.958429e-02  1.965311e-03
#> 
#> , , subj = 15
#> 
#>        par2
#> par1               Ka            Ke           V         Tlag1
#>   Ka     2.135819e-03  2.289351e-05 -0.12147786  1.116040e-03
#>   Ke     2.289351e-05  2.471571e-07 -0.00132394  8.168564e-06
#>   V     -1.214779e-01 -1.323940e-03  7.22340714 -1.262943e-02
#>   Tlag1  1.116040e-03  8.168564e-06 -0.01262943  9.092790e-03
#> 
#> , , subj = 16
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     3.263599e-04  5.282809e-06 -2.681502e-03 -3.757654e-04
#>   Ke     5.282809e-06  8.833559e-08 -4.632774e-05 -6.077011e-06
#>   V     -2.681502e-03 -4.632774e-05  2.508889e-02  3.080597e-03
#>   Tlag1 -3.757654e-04 -6.077011e-06  3.080597e-03  4.327545e-04
#> 
#> , , subj = 17
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     1.531043e-03  2.529613e-05 -0.0486476844 -0.0007883960
#>   Ke     2.529613e-05  4.179465e-07 -0.0008037645 -0.0000130260
#>   V     -4.864768e-02 -8.037645e-04  1.5457418769  0.0250506637
#>   Tlag1 -7.883960e-04 -1.302600e-05  0.0250506637  0.0004059771
#> 
#> , , subj = 18
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     2.905547e-04 -9.114520e-07  6.099822e-03  7.710707e-04
#>   Ke    -9.114520e-07  2.940902e-09 -1.921861e-05 -2.417231e-06
#>   V      6.099822e-03 -1.921861e-05  1.281454e-01  1.618592e-02
#>   Tlag1  7.710707e-04 -2.417231e-06  1.618592e-02  2.046297e-03
#> 
#> , , subj = 19
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     3.479309e-07 -3.975948e-08  5.964751e-05  1.095929e-07
#>   Ke    -3.975948e-08  4.544523e-09 -6.817964e-06 -1.250086e-08
#>   V      5.964751e-05 -6.817964e-06  1.022878e-02  1.875083e-05
#>   Tlag1  1.095929e-07 -1.250086e-08  1.875083e-05  3.539881e-08
#> 
#> , , subj = 20
#> 
#>        par2
#> par1               Ka            Ke             V         Tlag1
#>   Ka     2.069768e-09 -1.214542e-10  2.219400e-07  9.065267e-10
#>   Ke    -1.214542e-10  7.126941e-12 -1.302346e-08 -5.319506e-11
#>   V      2.219400e-07 -1.302346e-08  2.379850e-05  9.720634e-08
#>   Tlag1  9.065267e-10 -5.319506e-11  9.720634e-08  1.339727e-04
#> 
#> 
#> $postCor
#> , , subj = 1
#> 
#>        par2
#> par1    Ka Ke  V Tlag1
#>   Ka     1 -1  1     1
#>   Ke    -1  1 -1    -1
#>   V      1 -1  1     1
#>   Tlag1  1 -1  1     1
#> 
#> , , subj = 2
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000 -0.9805209  0.4641108 -0.8708237
#>   Ke    -0.9805209  1.0000000 -0.6290477  0.9504131
#>   V      0.4641108 -0.6290477  1.0000000 -0.8395569
#>   Tlag1 -0.8708237  0.9504131 -0.8395569  1.0000000
#> 
#> , , subj = 3
#> 
#>        par2
#> par1    Ka Ke  V Tlag1
#>   Ka     1 -1  1     1
#>   Ke    -1  1 -1    -1
#>   V      1 -1  1     1
#>   Tlag1  1 -1  1     1
#> 
#> , , subj = 4
#> 
#>        par2
#> par1    Ka Ke  V Tlag1
#>   Ka     1 -1  1     1
#>   Ke    -1  1 -1    -1
#>   V      1 -1  1     1
#>   Tlag1  1 -1  1     1
#> 
#> , , subj = 5
#> 
#>        par2
#> par1    Ka Ke  V Tlag1
#>   Ka     1 -1  1     1
#>   Ke    -1  1 -1    -1
#>   V      1 -1  1     1
#>   Tlag1  1 -1  1     1
#> 
#> , , subj = 6
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000  0.9227716 -0.9940664  0.9986908
#>   Ke     0.9227716  1.0000000 -0.9589949  0.9382127
#>   V     -0.9940664 -0.9589949  1.0000000 -0.9974327
#>   Tlag1  0.9986908  0.9382127 -0.9974327  1.0000000
#> 
#> , , subj = 7
#> 
#>        par2
#> par1              Ka           Ke            V        Tlag1
#>   Ka     1.000000000 -1.000000000  1.000000000  0.005628507
#>   Ke    -1.000000000  1.000000000 -1.000000000 -0.005628507
#>   V      1.000000000 -1.000000000  1.000000000  0.005628507
#>   Tlag1  0.005628507 -0.005628507  0.005628507  1.000000000
#> 
#> , , subj = 8
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000 -0.0423313 -0.6880271  0.9470787
#>   Ke    -0.0423313  1.0000000 -0.6958878  0.2804453
#>   V     -0.6880271 -0.6958878  1.0000000 -0.8844779
#>   Tlag1  0.9470787  0.2804453 -0.8844779  1.0000000
#> 
#> , , subj = 9
#> 
#>        par2
#> par1            Ka           Ke          V        Tlag1
#>   Ka     1.0000000 -0.065694600 -0.9271300  0.976961544
#>   Ke    -0.0656946  1.000000000 -0.2988724  0.006057307
#>   V     -0.9271300 -0.298872351  1.0000000 -0.919132806
#>   Tlag1  0.9769615  0.006057307 -0.9191328  1.000000000
#> 
#> , , subj = 10
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000 -0.9985041  0.9978038  0.8838258
#>   Ke    -0.9985041  1.0000000 -0.9998807 -0.8641770
#>   V      0.9978038 -0.9998807  1.0000000  0.8578806
#>   Tlag1  0.8838258 -0.8641770  0.8578806  1.0000000
#> 
#> , , subj = 11
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000  0.9032568 -0.7594297 -0.9323302
#>   Ke     0.9032568  1.0000000 -0.9647021 -0.6901675
#>   V     -0.7594297 -0.9647021  1.0000000  0.4754622
#>   Tlag1 -0.9323302 -0.6901675  0.4754622  1.0000000
#> 
#> , , subj = 12
#> 
#>        par2
#> par1            Ka          Ke          V       Tlag1
#>   Ka     1.0000000  0.97267331 -0.8071557  0.24745643
#>   Ke     0.9726733  1.00000000 -0.6500176  0.01905751
#>   V     -0.8071557 -0.65001765  1.0000000 -0.74180164
#>   Tlag1  0.2474564  0.01905751 -0.7418016  1.00000000
#> 
#> , , subj = 13
#> 
#>        par2
#> par1     Ka  Ke   V Tlag1
#>   Ka    NaN NaN NaN   NaN
#>   Ke    NaN NaN NaN   NaN
#>   V     NaN NaN NaN   NaN
#>   Tlag1 NaN NaN NaN     1
#> 
#> , , subj = 14
#> 
#>        par2
#> par1             Ka         Ke          V       Tlag1
#>   Ka     1.00000000 -0.7605865 -0.5035529  0.09066318
#>   Ke    -0.76058654  1.0000000 -0.1723504  0.56332693
#>   V     -0.50355289 -0.1723504  1.0000000 -0.89561837
#>   Tlag1  0.09066318  0.5633269 -0.8956184  1.00000000
#> 
#> , , subj = 15
#> 
#>        par2
#> par1            Ka         Ke           V       Tlag1
#>   Ka     1.0000000  0.9964226 -0.97801133  0.25324973
#>   Ke     0.9964226  1.0000000 -0.99085598  0.17231003
#>   V     -0.9780113 -0.9908560  1.00000000 -0.04927926
#>   Tlag1  0.2532497  0.1723100 -0.04927926  1.00000000
#> 
#> , , subj = 16
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000  0.9838947 -0.9371066 -0.9998793
#>   Ke     0.9838947  1.0000000 -0.9840848 -0.9828819
#>   V     -0.9371066 -0.9840848  1.0000000  0.9349176
#>   Tlag1 -0.9998793 -0.9828819  0.9349176  1.0000000
#> 
#> , , subj = 17
#> 
#>        par2
#> par1    Ka Ke  V Tlag1
#>   Ka     1  1 -1    -1
#>   Ke     1  1 -1    -1
#>   V     -1 -1  1     1
#>   Tlag1 -1 -1  1     1
#> 
#> , , subj = 18
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000 -0.9860062  0.9996588  0.9999907
#>   Ke    -0.9860062  1.0000000 -0.9899891 -0.9853568
#>   V      0.9996588 -0.9899891  1.0000000  0.9995435
#>   Tlag1  0.9999907 -0.9853568  0.9995435  1.0000000
#> 
#> , , subj = 19
#> 
#>        par2
#> par1            Ka         Ke          V      Tlag1
#>   Ka     1.0000000 -0.9998850  0.9998479  0.9875102
#>   Ke    -0.9998850  1.0000000 -0.9999970 -0.9856016
#>   V      0.9998479 -0.9999970  1.0000000  0.9854039
#>   Tlag1  0.9875102 -0.9856016  0.9854039  1.0000000
#> 
#> , , subj = 20
#> 
#>        par2
#> par1              Ka           Ke            V        Tlag1
#>   Ka     1.000000000 -1.000000000  1.000000000  0.001721518
#>   Ke    -1.000000000  1.000000000 -1.000000000 -0.001721518
#>   V      1.000000000 -1.000000000  1.000000000  0.001721518
#>   Tlag1  0.001721518 -0.001721518  0.001721518  1.000000000
#> 
#> 
#> $postMed
#>    id        Ka         Ke         V      Tlag1
#> 1   1 0.4320679 0.02425660  66.45764 0.54033955
#> 2   2 0.7438061 0.03914022 119.52708 0.02101887
#> 3   3 0.8960000 0.04307500 108.75000 2.10000000
#> 4   4 0.8959996 0.05594505 119.54995 0.69999785
#> 5   5 0.1040173 0.06782285 113.25195 0.02008672
#> 6   6 0.8959943 0.03515430  71.85063 1.97997170
#> 7   7 0.2160000 0.08366500  34.95000 1.81206790
#> 8   8 0.8959968 0.03515465  71.85031 1.97998410
#> 9   9 0.7919535 0.04406497 101.55520 0.89976788
#> 10 10 0.6560165 0.06188295  61.95186 0.82008218
#> 11 11 0.5840040 0.06782547  72.74957 1.33998020
#> 12 12 0.4641139 0.03021910  91.63719 1.01977840
#> 13 13 0.2160000 0.08366500  34.95000 1.81054350
#> 14 14 0.5520477 0.04406093 117.75206 0.29990845
#> 15 15 0.7999679 0.03416107  71.85822 1.05983960
#> 16 16 0.7519847 0.03515311  89.85172 0.70007653
#> 17 17 0.8959356 0.07375703  62.85724 1.18032190
#> 18 18 0.8959946 0.02327567  76.34939 1.77997280
#> 19 19 0.6560000 0.06188500  31.35000 1.94000010
#> 20 20 0.2160000 0.08366500  34.95000 1.81096390
#> 
#> $shrinkage
#>           Ka           Ke           V      Tlag1
#> 1 0.01242742 0.0005819863 0.001606317 0.00685773
#> 
#> $gridpts
#> [1] 20011
#> 
#> $nsub
#> [1] 20
#> 
#> $ab
#>       [,1]  [,2]
#> [1,] 1e-01   0.9
#> [2,] 1e-03   0.1
#> [3,] 3e+01 120.0
#> [4,] 0e+00   4.0
#> 
#> attr(,"class")
#> [1] "PMfinal" "NPAG"    "list"   
names(final)
#>  [1] "popPoints"  "popMean"    "popSD"      "popCV"      "popVar"    
#>  [6] "popCov"     "popCor"     "popMedian"  "popRanFix"  "postPoints"
#> [11] "postMean"   "postSD"     "postVar"    "postCov"    "postCor"   
#> [16] "postMed"    "shrinkage"  "gridpts"    "nsub"       "ab"        
final2 <- makeFinal(ITex$ITdata)
final2
#> $popMean
#>        Ka       Ke      V   Tlag1
#> 1 1.01972 0.045305 85.886 1.46089
#> 
#> $popSD
#>         Ka        Ke       V    Tlag1
#> 1 0.407229 0.0159415 37.8487 0.509122
#> 
#> $popCV
#>        Ka      Ke       V   Tlag1
#> 1 39.9355 35.1872 44.0685 34.8501
#> 
#> $popVar
#>          Ka           Ke        V     Tlag1
#> 1 0.1658355 0.0002541314 1432.524 0.2592052
#> 
#> $popCov
#>                 Ka            Ke            V         Tlag1
#> Ka     0.093286510 -0.0031803374    8.9118735 -4.455791e-02
#> Ke    -0.003180337  0.0002212130   -0.1991275  1.865823e-04
#> V      8.911873523 -0.1991274978 1274.5087774 -1.259831e+01
#> Tlag1 -0.044557910  0.0001865823  -12.5983075  2.140464e-01
#> 
#> $popCor
#>               Ka         Ke          V      Tlag1
#> Ka     1.0000000 -0.7000974  0.8173126 -0.3153272
#> Ke    -0.7000974  1.0000000 -0.3750200  0.0271151
#> V      0.8173126 -0.3750200  1.0000000 -0.7627581
#> Tlag1 -0.3153272  0.0271151 -0.7627581  1.0000000
#> 
#> $popMedian
#>        Ka        Ke       V   Tlag1
#> 1 1.04484 0.0428514 77.9248 1.53526
#> 
#> $postMean
#>    id       Ka        Ke        V    Tlag1
#> 1   1 1.129990 0.0205478  76.5444 1.626660
#> 2   2 1.357230 0.0324923 147.7150 0.543685
#> 3   3 1.418290 0.0450661  99.9225 2.074560
#> 4   4 1.150940 0.0518714 130.2580 0.766163
#> 5   5 1.591560 0.0438692 167.3070 0.850261
#> 6   6 1.006470 0.0389694  62.8183 1.938970
#> 7   7 0.963543 0.0488749  59.6357 2.030000
#> 8   8 1.032520 0.0391828  63.9842 1.982430
#> 9   9 1.145760 0.0417743 105.9730 1.232490
#> 10 10 0.858861 0.0537188  70.6835 1.310950
#> 11 11 0.849380 0.0624143  80.8237 1.475380
#> 12 12 1.354470 0.0262502 106.2990 1.595150
#> 13 13 0.257586 0.0744322  39.2744 1.610980
#> 14 14 1.059360 0.0418335 123.7810 0.716939
#> 15 15 1.057150 0.0311889  79.3053 1.459240
#> 16 16 1.108320 0.0344111  90.2250 1.379610
#> 17 17 0.755282 0.0700918  67.5731 1.180660
#> 18 18 1.013160 0.0282763  62.9503 1.791790
#> 19 19 0.570312 0.0628438  30.6302 1.901900
#> 20 20 0.714180 0.0579909  52.0175 1.750000
#> 
#> $postSD
#>    id       Ka         Ke        V     Tlag1
#> 1   1 0.299732 0.00565518 18.62770 0.2701600
#> 2   2 0.320341 0.00499317 22.73500 0.2422900
#> 3   3 0.268738 0.00596440 15.02050 0.0572197
#> 4   4 0.295727 0.00629333 17.69500 0.1231420
#> 5   5 0.320317 0.00563546 21.33880 0.0935929
#> 6   6 0.249506 0.00596204 10.74950 0.1009190
#> 7   7 0.271959 0.00579324  8.74851 0.3606920
#> 8   8 0.249261 0.00600871 10.91420 0.0930868
#> 9   9 0.291400 0.00581365 16.73990 0.3039630
#> 10 10 0.278485 0.00656339 10.61970 0.2915090
#> 11 11 0.278675 0.00654308 10.80940 0.2223330
#> 12 12 0.307533 0.00539695 21.19700 0.2763850
#> 13 13 0.100955 0.01292290  7.56560 0.1911950
#> 14 14 0.303632 0.00593398 19.01410 0.1738000
#> 15 15 0.290280 0.00597887 15.97080 0.2809980
#> 16 16 0.291584 0.00586331 16.60240 0.2864610
#> 17 17 0.272600 0.00678617  8.73285 0.3220030
#> 18 18 0.279528 0.00619569 13.94280 0.1848430
#> 19 19 0.226145 0.00762608  4.41272 0.1062260
#> 20 20 0.285015 0.00684034  8.17684 0.3595550
#> 
#> $postVar
#>    id         Ka           Ke         V       Tlag1
#> 1   1 0.08983927 3.198106e-05 346.99121 0.072986426
#> 2   2 0.10261836 2.493175e-05 516.88022 0.058704444
#> 3   3 0.07222011 3.557407e-05 225.61542 0.003274094
#> 4   4 0.08745446 3.960600e-05 313.11302 0.015163952
#> 5   5 0.10260298 3.175841e-05 455.34439 0.008759631
#> 6   6 0.06225324 3.554592e-05 115.55175 0.010184645
#> 7   7 0.07396170 3.356163e-05  76.53643 0.130098719
#> 8   8 0.06213105 3.610460e-05 119.11976 0.008665152
#> 9   9 0.08491396 3.379853e-05 280.22425 0.092393505
#> 10 10 0.07755390 4.307809e-05 112.77803 0.084977497
#> 11 11 0.07765976 4.281190e-05 116.84313 0.049431963
#> 12 12 0.09457655 2.912707e-05 449.31281 0.076388668
#> 13 13 0.01019191 1.670013e-04  57.23830 0.036555528
#> 14 14 0.09219239 3.521212e-05 361.53600 0.030206440
#> 15 15 0.08426248 3.574689e-05 255.06645 0.078959876
#> 16 16 0.08502123 3.437840e-05 275.63969 0.082059905
#> 17 17 0.07431076 4.605210e-05  76.26267 0.103685932
#> 18 18 0.07813590 3.838657e-05 194.40167 0.034166935
#> 19 19 0.05114156 5.815710e-05  19.47210 0.011283963
#> 20 20 0.08123355 4.679025e-05  66.86071 0.129279798
#> 
#> $postMed
#>    id        X1         X2        X3        X4
#> 1   1 1.1725606 0.02129534  74.06495 1.6276689
#> 2   2 1.3883977 0.03317120 144.45040 0.5498802
#> 3   3 1.4942498 0.04520522  99.14373 2.0754900
#> 4   4 1.1989619 0.05220644 128.90381 0.7751335
#> 5   5 1.6458539 0.04427601 165.24575 0.8559370
#> 6   6 1.0833952 0.03885700  62.78955 1.9417860
#> 7   7 1.0427318 0.04858538  59.78889 2.0300000
#> 8   8 1.1121369 0.03903037  64.02096 1.9842501
#> 9   9 1.2007115 0.04207777 104.85960 1.2410648
#> 10 10 0.9016617 0.05380257  70.32123 1.3160047
#> 11 11 0.9272884 0.06222125  80.73287 1.5113966
#> 12 12 1.4084108 0.02686801 103.76674 1.6116705
#> 13 13 0.2606130 0.07413613  39.42300 1.5823084
#> 14 14 1.1016885 0.04225238 122.26484 0.7220736
#> 15 15 1.1059247 0.03167587  77.88155 1.4623435
#> 16 16 1.1589821 0.03484994  88.82377 1.3861270
#> 17 17 0.7881036 0.06996770  67.45016 1.1811075
#> 18 18 1.0745367 0.02857761  62.14914 1.7942337
#> 19 19 0.6375351 0.06168130  31.13954 1.9110003
#> 20 20 0.7961764 0.05718801  52.64695 1.7500000
#> 
#> $shrinkage
#>          Ka        Ke         V     Tlag1
#> 1 0.4656046 0.1730608 0.1547893 0.2155101
#> 
#> $nsub
#> [1] 20
#> 
#> $ab
#>       [,1]        [,2]
#> [1,] 1e-08   3.0558655
#> [2,] 1e-08   0.1250127
#> [3,] 1e-08 275.1293673
#> [4,] 1e-08   4.0065024
#> 
#> attr(,"class")
#> [1] "PMfinal" "IT2B"    "list"   
names(final2)
#>  [1] "popMean"   "popSD"     "popCV"     "popVar"    "popCov"    "popCor"   
#>  [7] "popMedian" "postMean"  "postSD"    "postVar"   "postMed"   "shrinkage"
#> [13] "nsub"      "ab"