Generated observed vs. predicted data

Generates an observed vs. predicted data.frame from an NPAG or IT2B object

Usage

makeOP(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeOP is a data frame of class PMop, which has a population and posterior prediction object (also class PMop) for each output equation. Each of these has 13 columns:

id

subject identification

time

observation time in relative hours

obs

observation

pred

prediction

pred.type

Population predictions based on Bayesian prior parameter value distribution, or individual predictions based on Bayesian posterior parameter value distributions

icen

Predictions based on mean or median of Bayesian pred.type parameter values

outeq

output equation number

block

dosing block number for each subject, as defined by dose resets (evid=4).

obsSD

standard deviation of the observation based on the assay error polynomial

d

prediction error, pred-obs

ds

squared prediction error

wd

weighted prediction error, which is the prediction error divided by the obsSD

wds

weighted squared prediction error

A plot method exists in plot for PMop objects.

Details

makeOP will parse the output of NPparse or ITparse to generate a data.frame suitable for analysis and plotting of observed vs. population or or individual predicted outputs.

See also

Author

Michael Neely

Examples

op <- makeOP(NPex$NPdata)
op
#>    id   time   obs     pred pred.type icen outeq block     obsSD            d
#> 1   1 120.00 10.44 3.663302       pop mean     1     1 1.2565350  -6.77669827
#> 2   1 121.00 12.89 3.479807       pop mean     1     1 1.5248181  -9.41019288
#> 3   1 122.00 14.98 6.326221       pop mean     1     1 1.7490961  -8.65377937
#> 4   1 125.99 16.69 8.846097       pop mean     1     1 1.9294571  -7.84390347
#> 5   1 129.00 20.15 7.868253       pop mean     1     1 2.2857564 -12.28174733
#> 6   1 132.00 14.97 6.787836       pop mean     1     1 1.7480330  -8.18216416
#> 7   1 143.98 12.57 3.672533       pop mean     1     1 1.4901063  -8.89746689
#> 8   2 120.00  3.56 3.663302       pop mean     1     1 0.4721422   0.10330173
#> 9   2 120.98  5.84 3.483385       pop mean     1     1 0.7371530  -2.35661463
#> 10  2 121.98  6.54 6.274216       pop mean     1     1 0.8175083  -0.26578435
#> 11  2 126.00  6.14 8.843913       pop mean     1     1 0.7716490   2.70391316
#> 12  2 129.02  6.56 7.860880       pop mean     1     1 0.8197972   1.30088042
#> 13  2 132.02  4.44 6.780971       pop mean     1     1 0.5750222   2.34097097
#> 14  2 144.00  3.76 3.668761       pop mean     1     1 0.4955897  -0.09123942
#> 15  3 120.08  4.06 3.648273       pop mean     1     1 0.5306885  -0.41172728
#> 16  3 121.07  3.24 3.467312       pop mean     1     1 0.4345457   0.22731214
#> 17  3 122.08  3.09 6.526507       pop mean     1     1 0.4168884   3.43650685
#> 18  3 126.08  7.98 8.825962       pop mean     1     1 0.9813218   0.84596248
#> 19  3 129.05  7.23 7.849820       pop mean     1     1 0.8962522   0.61981964
#> 20  3 132.10  4.71 6.753569       pop mean     1     1 0.6064377   2.04356863
#> 21  3 144.08  3.82 3.653709       pop mean     1     1 0.5026164  -0.16629083
#> 22  4 120.00  2.10 3.663302       pop mean     1     1 0.2998044   1.56330173
#> 23  4 121.00  3.05 3.479807       pop mean     1     1 0.4121760   0.42980712
#> 24  4 122.02  5.21 6.377442       pop mean     1     1 0.6644285   1.16744239
#> 25  4 126.00  5.09 8.843913       pop mean     1     1 0.6505328   3.75391316
#> 26  4 129.03  4.24 7.857194       pop mean     1     1 0.5517061   3.61719380
#> 27  4 132.00  3.69 6.787836       pop mean     1     1 0.4873875   3.09783584
#> 28  4 144.02  1.96 3.664992       pop mean     1     1 0.2831706   1.70499193
#> 29  5 120.00  2.93 3.663302       pop mean     1     1 0.3980298   0.73330173
#> 30  5 121.00  2.64 3.479807       pop mean     1     1 0.3637857   0.83980712
#> 31  5 122.00  4.80 6.326221       pop mean     1     1 0.6168939   1.52622063
#> 32  5 126.00  3.70 8.843913       pop mean     1     1 0.4885595   5.14391316
#> 33  5 129.02  4.13 7.860880       pop mean     1     1 0.5388658   3.73088042
#> 34  5 132.00  2.81 6.787836       pop mean     1     1 0.3838697   3.97783584
#> 35  5 144.00  2.21 3.668761       pop mean     1     1 0.3128605   1.45876058
#> 36  6 120.00  6.92 3.663302       pop mean     1     1 0.8609314  -3.25669827
#> 37  6 121.00  6.89 3.479807       pop mean     1     1 0.8575083  -3.41019288
#> 38  6 121.98  6.64 6.274216       pop mean     1     1 0.8289489  -0.36578435
#> 39  6 126.00 13.72 8.843913       pop mean     1     1 1.6143907  -4.87608684
#> 40  6 129.00 12.69 7.868253       pop mean     1     1 1.5031348  -4.82174733
#> 41  6 131.98 10.58 6.794706       pop mean     1     1 1.2720217  -3.78529359
#> 42  6 144.98  6.62 3.488576       pop mean     1     1 0.8266616  -3.13142389
#> 43  7 120.00  5.41 3.663302       pop mean     1     1 0.6875572  -1.74669827
#> 44  7 121.03  4.46 3.474447       pop mean     1     1 0.5773517  -0.98555338
#> 45  7 122.03  4.54 6.402763       pop mean     1     1 0.5866658   1.86276266
#> 46  7 126.02 12.19 8.839506       pop mean     1     1 1.4487576  -3.35049448
#> 47  7 129.08 12.10 7.838756       pop mean     1     1 1.4389440  -4.26124367
#> 48  7 132.03  8.61 6.777541       pop mean     1     1 1.0523602  -1.83245933
#> 49  7 144.03  6.37 3.663109       pop mean     1     1 0.7980369  -2.70689095
#> 50  8 120.00  6.19 3.663302       pop mean     1     1 0.7773898  -2.52669827
#> 51  8 121.03  6.33 3.474447       pop mean     1     1 0.7934514  -2.85555338
#> 52  8 122.00  6.24 6.326221       pop mean     1     1 0.7831283   0.08622063
#> 53  8 125.98 13.03 8.848266       pop mean     1     1 1.5399734  -4.18173386
#> 54  8 128.98 11.86 7.875624       pop mean     1     1 1.4127363  -3.98437649
#> 55  8 132.00 11.45 6.787836       pop mean     1     1 1.3678360  -4.66216416
#> 56  8 143.98  7.83 3.672533       pop mean     1     1 0.9643513  -4.15746689
#> 57  9 120.00  2.85 3.663302       pop mean     1     1 0.3885913   0.81330173
#> 58  9 120.97  3.70 3.485176       pop mean     1     1 0.4885595  -0.21482412
#> 59  9 122.00  6.65 6.326221       pop mean     1     1 0.8300925  -0.32377937
#> 60  9 125.98  6.81 8.848266       pop mean     1     1 0.8483759   2.03826614
#> 61  9 128.98  6.51 7.875624       pop mean     1     1 0.8140742   1.36562351
#> 62  9 132.00  7.48 6.787836       pop mean     1     1 0.9246691  -0.69216416
#> 63  9 143.98  4.51 3.672533       pop mean     1     1 0.5831738  -0.83746689
#> 64 10 120.00  2.93 3.663302       pop mean     1     1 0.3980298   0.73330173
#> 65 10 121.00  4.36 3.479807       pop mean     1     1 0.5657004  -0.88019288
#> 66 10 122.02  7.79 6.377442       pop mean     1     1 0.9598222  -1.41255761
#> 67 10 126.00 11.02 8.843913       pop mean     1     1 1.3205709  -2.17608684
#> 68 10 129.00  8.86 7.868253       pop mean     1     1 1.0804437  -0.99174733
#> 69 10 131.97  6.09 6.798144       pop mean     1     1 0.7659057   0.70814382
#> 70 10 144.00  4.15 3.668761       pop mean     1     1 0.5412012  -0.48123942
#> 71 11 120.00  2.09 3.663302       pop mean     1     1 0.2986169   1.57330173
#> 72 11 121.03  2.68 3.474447       pop mean     1     1 0.3685139   0.79444662
#> 73 11 122.00  4.71 6.326221       pop mean     1     1 0.6064377   1.61622063
#> 74 11 125.98  7.71 8.848266       pop mean     1     1 0.9507593   1.13826614
#> 75 11 129.00  6.31 7.868253       pop mean     1     1 0.7911580   1.55825267
#> 76 11 132.00  5.82 6.787836       pop mean     1     1 0.7348502   0.96783584
#>              ds         wd          wds
#> 1  4.592364e+01 -5.3931631  29.08620853
#> 2  8.855173e+01 -6.1713544  38.08561523
#> 3  7.488790e+01 -4.9475723  24.47847148
#> 4  6.152682e+01 -4.0653422  16.52700721
#> 5  1.508413e+02 -5.3731654  28.87090620
#> 6  6.694781e+01 -4.6807835  21.90973378
#> 7  7.916492e+01 -5.9710282  35.65317812
#> 8  1.067125e-02  0.2187937   0.04787068
#> 9  5.553633e+00 -3.1969137  10.22025748
#> 10 7.064132e-02 -0.3251152   0.10569988
#> 11 7.311146e+00  3.5040716  12.27851756
#> 12 1.692290e+00  1.5868320   2.51803592
#> 13 5.480145e+00  4.0710964  16.57382577
#> 14 8.324632e-03 -0.1841027   0.03389382
#> 15 1.695194e-01 -0.7758360   0.60192150
#> 16 5.167081e-02  0.5231029   0.27363666
#> 17 1.180958e+01  8.2432307  67.95085210
#> 18 7.156525e-01  0.8620643   0.74315489
#> 19 3.841764e-01  0.6915683   0.47826673
#> 20 4.176173e+00  3.3697915  11.35549450
#> 21 2.765264e-02 -0.3308504   0.10946196
#> 22 2.443912e+00  5.2144063  27.19003265
#> 23 1.847342e-01  1.0427756   1.08738093
#> 24 1.362922e+00  1.7570624   3.08726838
#> 25 1.409186e+01  5.7705214  33.29891673
#> 26 1.308409e+01  6.5563780  42.98609233
#> 27 9.596587e+00  6.3560022  40.39876362
#> 28 2.906997e+00  6.0210767  36.25336503
#> 29 5.377314e-01  1.8423285   3.39417441
#> 30 7.052760e-01  2.3085212   5.32927018
#> 31 2.329349e+00  2.4740408   6.12087784
#> 32 2.645984e+01 10.5287340 110.85424038
#> 33 1.391947e+01  6.9235803  47.93596476
#> 34 1.582318e+01 10.3624630 107.38063931
#> 35 2.127982e+00  4.6626557  21.74035787
#> 36 1.060608e+01 -3.7827616  14.30928501
#> 37 1.162942e+01 -3.9768626  15.81543630
#> 38 1.337982e-01 -0.4412628   0.19471288
#> 39 2.377622e+01 -3.0203884   9.12274617
#> 40 2.324925e+01 -3.2077943  10.28994425
#> 41 1.432845e+01 -2.9758090   8.85543903
#> 42 9.805816e+00 -3.7880361  14.34921753
#> 43 3.050955e+00 -2.5404406   6.45383855
#> 44 9.713155e-01 -1.7070242   2.91393164
#> 45 3.469885e+00  3.1751682  10.08169340
#> 46 1.122581e+01 -2.3126674   5.34843070
#> 47 1.815820e+01 -2.9613686   8.76970410
#> 48 3.357907e+00 -1.7412853   3.03207465
#> 49 7.327259e+00 -3.3919369  11.50523582
#> 50 6.384204e+00 -3.2502333  10.56401634
#> 51 8.154185e+00 -3.5989014  12.95209145
#> 52 7.433998e-03  0.1100977   0.01212151
#> 53 1.748690e+01 -2.7154585   7.37371497
#> 54 1.587526e+01 -2.8203258   7.95423737
#> 55 2.173577e+01 -3.4084235  11.61735047
#> 56 1.728453e+01 -4.3111538  18.58604677
#> 57 6.614597e-01  2.0929488   4.38043484
#> 58 4.614940e-02 -0.4397092   0.19334420
#> 59 1.048331e-01 -0.3900522   0.15214069
#> 60 4.154529e+00  2.4025507   5.77224997
#> 61 1.864928e+00  1.6775173   2.81406426
#> 62 4.790912e-01 -0.7485533   0.56033209
#> 63 7.013508e-01 -1.4360504   2.06224068
#> 64 5.377314e-01  1.8423285   3.39417441
#> 65 7.747395e-01 -1.5559346   2.42093258
#> 66 1.995319e+00 -1.4716867   2.16586180
#> 67 4.735354e+00 -1.6478380   2.71536994
#> 68 9.835628e-01 -0.9179074   0.84255407
#> 69 5.014677e-01  0.9245836   0.85485484
#> 70 2.315914e-01 -0.8892061   0.79068740
#> 71 2.475278e+00  5.2686299  27.75846108
#> 72 6.311454e-01  2.1558118   4.64752457
#> 73 2.612169e+00  2.6651057   7.10278864
#> 74 1.295650e+00  1.1972179   1.43333080
#> 75 2.428151e+00  1.9695845   3.87926328
#> 76 9.367062e-01  1.3170519   1.73462566
#>  [ reached 'max' / getOption("max.print") -- omitted 480 rows ]
names(op)
#>  [1] "id"        "time"      "obs"       "pred"      "pred.type" "icen"     
#>  [7] "outeq"     "block"     "obsSD"     "d"         "ds"        "wd"       
#> [13] "wds"