Plot Pmetrics Multiple-Model Optimal Sampling Objects

Plots PM_opt objects

Usage

# S3 method for class 'PM_opt'
plot(x, line = list(probs = NA), times = T, print = TRUE, ...)

Arguments

x

A PM_opt object

line

Passed to plot.PM_sim with default as list(probs = NA).

times

Format the vertical lines for optimal times. Default is dashed red line. r template("line")`

print

If TRUE, will print the plotly object and return it. If FALSE, will only return the plotly object.

...

Other parameters to pass to plot.PM_sim.

probs

Default is NA. See plot.PM_sim for details.

Value

Plots the simulation profiles with MM optimal times indicated as vertical lines.

Details

Simulated observations are plotted on the y-axis vs. time on the x.axis. Optimal sampling times are indicated as vertical lines. Defaults for optimal sample times are red, dash, width 2. Defaults for the line format are as for plot.PM_sim.

See also

Author

Michael Neely