Plot NPAG Cycle Information

Plot PMcycle objects

Usage

# S3 method for PMcycle
plot(data, x.leg = 0, y.leg = 1, cex.leg = 1.2, omit, col, out = NA, ...)

Arguments

data

The name of an PMcycle data object generated by makeCycle

x.leg

Porportionate location along the X-axis to place legend; 0 (default) is at left, 1 at right.

y.leg

Porportionate location along the X-axis to place legend; 0 is at bottom, 1 (default) at top.

cex.leg

Porportionate size of legend text.

omit

Deceimal between 0 and 1 specifying the proportion of “burn-in” cycles to omit from the plots. If missing, the first 20% will be omitted.

col

A vector of colors for the curves, which will be recycled if too short. Not mandatory.

out

Direct output to a PDF, EPS or image file. Format is a named list whose first argument, type is one of the following character vectors: “pdf”, “eps” (maps to postscript), “png”, “tiff”, “jpeg”, or “bmp”. Other named items in the list are the arguments to each graphic device. PDF and EPS are vector images acceptable to most journals in a very small file size, with scalable (i.e. infinite) resolution. The others are raster images which may be very large files at publication quality dots per inch (DPI), e.g. 800 or 1200. Default value is NA which means the output will go to the current graphic device (usually the monitor). For example, to output an eps file, out=list(“eps”) will generate a 7x7 inch (default) graphic.

...

Additional R plotting parameters.

Value

Plots a panel with the following windows: -2 times the log-likelihood at each cycle, gamma/lambda at each cycle; Akaike Information Criterion at each cyle and Bayesian (Schwartz) Information Criterion at each cycle, the mean parameter values at each cycle (normalized to starting values); the normalized standard deviation of the population distribution for each parameter at each cycle; and the normalized median parameter values at each cycle.

See also

Author

Michael Neely

Examples

data(cycle.1)
#> Warning: data set ‘cycle.1’ not found
plot(cycle.1)
#> Error in plot(cycle.1): object 'cycle.1' not found
plot(cycle.1,omit=0)
#> Error in plot(cycle.1, omit = 0): object 'cycle.1' not found