Extract CDC pediatric BMI z-scores

Will extract BMI z-scores based on a single age in months and sex. Overweight is a z-score of >1.04, and obese is a z-score > 1.64. Calculations are based on CDC formulae

For a z-score > 3, which indicates an extreme BMI, consider using the modified z-score and percentile.

Usage

zBMI(agemos, sex, bmi, wt, ht, data = "CDC")

Arguments

agemos

The age in months. Should be between 24 and 240.5.

sex

A single quoted character: "M" for males, "F" for females. Default is "M".

bmi

The individual's BMI. If specified, wt and ht are not necessary and will be ignored.

wt

The individual's weight in kg as an alternative to specifying BMI. Will be ignored if BMI is specified.

ht

The individual's height in centimeters. Required if wt is specified and BMI is not. Ignored if BMI is specified.

data

Source data for calculations. Default is "CDC" which uses the cdc_bmi dataset. The alternative is "NHANES", which uses the ger_bmi dataset.

Value

A list with objects calculated for agemos and sex.

z Z-score

mod_z Modified Z-score for extreme BMI

per BMI percentile

mod_per Modified BMI percentile

Author

Michael Neely

Examples

zBMI(agemos = 36, bmi = 15)
#> $z
#> [1] -0.9308131
#> 
#> $mod_z
#> [1] -0.7077054
#> 
#> $per
#> [1] 0.1759751
#> 
#> $mod_per
#> [1] 0.2395641
#>