Defines the PM_data object

PM_data R6 objects containing raw, standardized and valid data, and methods to process the data

Details

PM_data objects are passed to PM_fit objects to initiate a population analysis. The object is created by reading a delimited file in the current working directory. The data will be transformed into the standard format which is the same for all engines, with a report of any assumptions that were necessary to standardize the data. PMcheck is called on the standard data to evaluate for errors. There are a number of methods defined for a PM_data object, including to write the standard data back to a file for future use, to summarize and to plot the object, and to conduct a non-compartmental analysis on the raw data using makeNCA.

Public fields

data

Data frame containing the data to be modeled

standard_data

Data frame containing standardized version of the data

Methods

Public methods

PM_data$new()

PM_data$write()

PM_data$nca()

PM_data$plot()

PM_data$print()

PM_data$summary()

PM_data$clone()

Method new()

Create new data object

Usage

PM_data$new(data, dt = c("Ymd", "HS"), quiet = F)

Arguments

data

A quoted name of a file with full path if not in the working directory, or an unquoted name of a data frame in the current R environment.

dt

Date time format as a character vector whose first element is date format and second is time. Default is c("Ymd", "HS"). Use the following abbreviations:

Y = 4 digit year

y = 2 digit year

m = decimal month (1, 2, ..., 12)

d = decimal day (1, 2, ..., 31)

H = hours (0-23)

M = minutes (0-59) Other formats are possible. See strptime for these, omitting the % before each abbreviation.

quiet

Quietly validate. Default is FALSE.

Details

Creation of a new PM_data objects from a file or a data frame. Data will be standardized and checked automatically to a fully specified, valid data object.

Method write()

Write data to file

Usage

PM_data$write(file_name)

Arguments

file_name

A quoted name of the file to create with full path if not in the working directory.

Details

Writes a delimited file (e.g. comma-separated) from the standard_data field

Method nca()

Perform non-compartmental analysis

Usage

PM_data$nca(...)

Arguments

...

Arguments passed to makeNCA.

Details

See makeNCA.

Method plot()

Plot method

Usage

PM_data$plot(...)

Arguments

...

Arguments passed to plot.PM_data

Details

See plot.PMmatrix.

Method print()

Print method

Usage

PM_data$print(standard = F, viewer = T, ...)

Arguments

standard

Display the standardized data if TRUE. Default is FALSE.

viewer

Display the Viewer if TRUE. Default is TRUE.

...

Other arguments to print.data.frame. Only passed if viewer = FALSE.

Details

Displays the PM_data object in a variety of ways.

Method summary()

Summary method

Usage

PM_data$summary(...)

Arguments

...

Arguments passed to summary.PMmatrix.

Details

See summary.PMmatrix.

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.