Final Cycle Population Values

Final Cycle Population Values

Final Cycle Population Values

Details

Contains final cycle information from run.

Contains the results of makeFinal, which is a list suitable for analysis and plotting of final cycle population values.

However, if you wish to manipulate the entire data frame, use the data field, e.g. probs <- run1$final$data$popPoints %>% select(prob). This will select the probabilities of the support points. If you are unfamiliar with the %>% pipe function, please type help("%>%", "magrittr") into the R console and look online for instructions/tutorials in tidyverse, a powerful approach to data manipulation upon which Pmetrics is built.

Creation of new PM_final object is automatic and not generally necessary for the user to do.

See summary.PMfinal.

See plot.PMfinal.

Public fields

popPoints

(NPAG only) Data frame of the final cycle joint population density of grid points with column names equal to the name of each random parameter plus prob for the associated probability of that point

popMean

The final cycle mean for each random parameter distribution

popSD

The final cycle standard deviation for each random parameter distribution

popCV

The final cycle coefficient of variation (SD/Mean) for each random parameter distribution

popVar

The final cycle variance for each random parameter distribution

popCov

The final cycle random parameter covariance matrix

popCor

The final cycle random parameter correlation matrix

popMedian

The final cycle median values for each random parameter, i.e. those that have unknown mean and unknown variance, both of which are fitted during the run

popRanFix

The final cycle median values for each parameter that is random but fixed to be the same for all subjects, i.e. unknown mean, zero variance, with only mean fitted in the run

postPoints

(NPAG only) Data frame of posterior population points for each of the first 100 subject, with columns id, point, parameters and probability. The first column is the subject, the second column has the population point number, followed by the values for the parameters in that point and the probability.

postMean

A nsub x npar data frame containing the means of the posterior distributions for each parameter.

postSD

A nsub x npar data frame containing the SDs of the posterior distributions for each parameter.

postVar

A nsub x npar data frame containing the variances of the posterior distributions for each parameter.

postCov

NPAG only: An array of dimensions npar x npar x nsub that contains the covariances of the posterior distributions for each parameter and subject.*

postCor

NPAG only: An array of dimensions npar x npar x nsub that contains the correlations of the posterior distributions for each parameter and subject.

postMed

A nsub x npar data frame containing the medians of the posterior distributions for each parameter.*

shrinkage

A data frame with the shrinkage for each parameter. popVar is comprised of variance(EBE) + variance(EBD), where EBE is the Empirical Bayes Estimate or mean of the posterior distribution for the parameter. EBD is the Empirical Bayes Distribution, or the full Bayesian posterior distribution. In other words, if Bayesian posterior distributions are wide for a given parameter due to sparse or uninformative sampling, then most of the population variance is due to this variance and shrinkage of the EBE variance is high because individual posterior estimates shrink towards the population mean.

gridpts

(NPAG only) Initial number of support points

nsub

Number of subjects

ab

Matrix of boundaries for random parameter values

data

A data frame combining all the above fields as its columns

Methods

Public methods

PM_final$new()

PM_final$summary()

PM_final$plot()

PM_final$clone()

Method new()

Create new object populated with final cycle information

Usage

PM_final$new(final)

Arguments

final

The parsed output from makeFinal.

Method summary()

Summary method

Usage

PM_final$summary(...)

Arguments

...

Arguments passed to summary.PMfinal

Method plot()

Plot method

Usage

PM_final$plot(...)

Arguments

...

Arguments passed to plot.PMfinal

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_final$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.