Defines the PM_model class

PM_model objects contain the variables, covariates, equations and error models necessary to run a population analysis.

Details

PM_model objects are passed to PM_fit objects to initiate a population analysis. The object is created by defining a list of lists directly in R, or by reading a model text file in the current working directory.

Methods

Public methods

PM_model$new()

PM_model$print()

PM_model$update()

PM_model$write()

PM_model$clone()

Method new()

Build a new PM_model from a variety of inputs.

Usage

PM_model$new(model, ..., julia = F)

Arguments

model

This can be a quoted name of a model text file in the working directory which will be read and passed to Fortran engines unless julia = TRUE in which case is will be passed as a Julia model. It can be a list of lists that defines the model directly in R. Similarly, it can be a function that defines a Julia model directly in R. See the user manual for more help on directly defining models in R.

...

Additional arguments passed to creation of new Julia model.

julia

Controls whether a filename defines a Fortran or Julia model. Default is FALSE, i.e. using Fortran.

Method print()

Print a model object to the console in readable format

Usage

PM_model$print(...)

Arguments

...

Not used currently.

Method update()

Update selected elements of a model object

Usage

PM_model$update(changes_list)

Arguments

changes_list

The named list containing elements and values to update.

Examples

\dontrun{mod1$update(list(
  pri = list(
  )
)) 
}

Method write()

Write a PM_model object to a text file

Usage

PM_model$write(model_path = "genmodel.txt", engine = "npag")

Arguments

model_path

Full name of the file to be created, including the path relative to the current working directory

engine

Currently only "npag".

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_model$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `PM_model$update`
## ------------------------------------------------

if (FALSE) mod1$update(list(
  pri = list(
  )
))