Observed vs. predicted data

Observed vs. predicted data

Observed vs. predicted data

Details

Contains observed vs. predicted data after a run

Contains the results of makeOP, which is a data frame suitable for analysis and plotting of observed vs. population or or individual predicted outputs. To provide a more traditional experience in R, the data frame is separated by columns into fields, e.g. id or time. This allows you to access them in an S3 way, e.g. run1$op$time if run1 is a PM_result object.

However, if you wish to manipulate the entire data frame, use the data field, e.g. trough <- run1$op$data %>% filter(time == 24). If you are unfamiliar with the %>% pipe function, please type help("%>%", "magrittr") into the R console and look online for instructions/tutorials in tidyverse, a powerful approach to data manipulation upon which Pmetrics is built.

Creation of new PM_op object is automatic and not generally necessary for the user to do.

See plot.PM_op.

See summary.PMop.

See makeAUC

Public fields

id

subject identification

time

observation time in relative units, usually hours

obs

observation

pred

prediction

pred.type

Population predictions based on Bayesian prior parameter value distribution, or individual predictions based on Bayesian posterior parameter value distributions

icen

Predictions based on mean or median of Bayesian pred.typeparameter values

outeq

output equation number

block

dosing block number for each subject, as defined by dose resets (evid=4).

obsSD

standard deviation of the observation based on the assay error polynomial

d

prediction error, pred - obs

ds

squared prediction error

wd

weighted prediction error, which is the prediction error divided by the obsSD

wds

weighted squared prediction error

data

A data frame combining all the above fields as its columns

Methods

Public methods

PM_op$new()

PM_op$plot()

PM_op$summary()

PM_op$auc()

PM_op$clone()

Method new()

Create new object populated with observed vs. predicted data

Usage

PM_op$new(op)

Arguments

op

The parsed output from makeOP.

Method plot()

Plot method

Usage

PM_op$plot(...)

Arguments

...

Arguments passed to plot.PM_op

Method summary()

Summary method

Usage

PM_op$summary(...)

Arguments

...

Arguments passed to summary.PMop

Method auc()

Calculate AUC

Usage

PM_op$auc(...)

Arguments

...

Arguments passed to makeAUC

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_op$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.