Define object for PTA

If the $save method has previously been invoked on a PM_pta object, this function will load those results.

Arguments

file_name

The name of the .rds file to load.

Value

A PM_pta object

Details

This object class contains results of simulations and a probability of target attainment analysis.

Creating this object first requires a simulation of appropriate regimens from a model using the $sim method for a PM_result object. That simulation result is passed to the makePTA function when a new PM_pta is created.

The saved object is an .rds file. When loaded, it should be assigned to an R object, e.g. pta1 <- PM_pta$load("filename").

Author

Julian Otalvaro and Michael Neely

Public fields

results

Contains the raw results. See makePTA.

outcome

Contains summary of raw results. See makePTA.

Methods

Public methods

PM_pta$new()

PM_pta$save()

PM_pta$summary()

PM_pta$plot()

PM_pta$load()

PM_pta$clone()

Method new()

Create a new PM_pta object

Usage

PM_pta$new(simdata, targets, target.type, ...)

Arguments

simdata

Output of $sim method for PM_result object

targets

See makePTA.

target.type

See makePTA.

...

Other arguments passed to makePTA.

Method save()

Save the current PM_pta object into a .rds file.

Usage

PM_pta$save(file_name = "PMpta.rds")

Arguments

file_name

Name of the file to be created, the default is PMpta.rds

Method summary()

Summarize the PM_pta object. See summary.PMpta.

Usage

PM_pta$summary(...)

Arguments

...

Arguments passed to summary.PMpta

Method plot()

Plot the PM_pta object. See plot.PM_pta.

Usage

PM_pta$plot(...)

Arguments

...

Arguments passed to plot.PM_pta

Method load()

Returns a PM_pta object based on the information found in a specified rds file.

Usage

PM_pta$load(file_name)

Arguments

file_name

Name of the file to be read, the default is PMpta.rds

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_pta$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.