Generate a report

Generates a report from a specified Rmd template

Usage

PM_report(x, template, path, show = TRUE, quiet = TRUE)

Arguments

x

A PM_result object obtained from PM_load.

template

If missing, the default Pmetrics report template as specified in getPMoptions is used. It can be changed with setPMoptions. Otherwise, the value for template can be "plotly", "ggplot", or "none".

path

The path for the generated report, defaults to a temporary folder.

show

Controls if the report should be automatically opened on generation, defaults to TRUE

quiet

If TRUE (default), suppresses knitr output about report generation. Progress messages will still be displayed.

Value

Generates an HTML-report in the folder specified by path.

See also

Author

Markus Hovd, Julian Otalvaro, and Michael Neely