Pmetrics validation object

Contains results of internal validation by simulation to permit generation of visual predictive checks (VPCs), prediction corrected visual predictive checks, (pcVPCs), normalized prediction distribution errors (NPDE), and numerical predictive checks.

Details

This object is created by running the make_valid method in a PM_result object. It contains all the information necessary to internally validate the result by simulation methods.

See also

Public fields

simdata

Simulated data created in the validation process

timeBinMedian

Median times for cluster bins

tadBinMedian

Median times after previous doses for cluster bins

opDF

Observed-predicted data frame

npde

Data for Normalized Prediction Distribution Error

npde_tad

Data for Normalized Prediction Distribution Error using Time After Dose if available

Methods

Public methods

PM_valid$new()

PM_valid$plot()

PM_valid$clone()

Method new()

Create a new PM_valid object from a PM_result

Usage

PM_valid$new(result, ...)

Arguments

result

The PM_result object

...

Additional arguments to ultimately pass to makeValid

Method plot()

Plot method. Calls plot.PMvalid.

Usage

PM_valid$plot(...)

Arguments

...

Arguments to pass to [plot.PMvalid].

Method clone()

The objects of this class are cloneable with this method.

Usage

PM_valid$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.