Calculation of AUCs

Calculates AUC from a variety of inputs

Usage

makeAUC(
  data,
  formula,
  include,
  exclude,
  start = 0,
  end = Inf,
  icen = "median",
  outeq = 1,
  block = 1,
  method = "linear",
  addZero = F
)

Arguments

data

A suitable data object of the PMpop class (see makePop), PMpost class (see makePost), PMop class (see makeOP), the PMsim class (see SIMparse), or some other suitable dataframe with at least time/observation columns referred to by formula, with an “id” column (so named) if necessary.

formula

A formula of the form obs ~ time. This is only required with data that is not of class PMpop, PMpost, PMop or PMsim.

include

A vector of subject IDs to include in the AUC calculations, e.g. c(1:3,5,15)

exclude

A vector of subject IDs to exclude in the AUC calculations, e.g. c(4,6:14,16:20)

start

Specify the time to begin AUC calculations. Default is 0.

end

Specify the time to end AUC calculations so that AUC is calculated from start to end. Default for end is the maximum observation time for each subject. Subjects with insufficient data for a specified interval will have AUC calculated for the available data, not to exceed the specified interval.

icen

Only relevant for PMpost or PMpop objects which have predictions based on median or mean of each subject's Bayesian posterior parameter distribution. Default is "median", but could be "mean".

outeq

Specify which output equation is to be used. Default is 1.

block

Specify which observation block (separated by EVID=4) is to be used for each subject. Default is 1.

method

Default is "linear" for AUC trapezoidal calculation. Any other value will result in linear up, log down.

addZero

Boolean to add a zero concentration at time 0. Default is FALSE.

Value

The output of makeAUC is a dataframe of class PMauc, which has 2 columns:

id

subject identification

tau

AUC from start to end

Details

makeAUC will calculate the area under the time concentration curve using the trapezoidal approximation from a variety of inputs. If a PMpost, PMop, or PMsim object is specified, formula is not required.

See also

Author

Michael Neely

Examples

data(NPdata.1)
#> Warning: data set ‘NPdata.1’ not found
op <- makeOP(NPdata.1)
#> Error in inherits(data, "NPAG"): object 'NPdata.1' not found
makeAUC(op)
#> Error in inherits(data, c("PMsim", "PM_sim", "PMop", "PM_op", "PMpop",     "PM_pop", "PMpost", "PM_post"), which = T): object 'op' not found