Extract covariate data

Generates an data.frame with subject-specific covariate data from an NPAG or IT2B object

Usage

makeCov(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeCov is a dataframe of class PMcov, which has the following columns:

id

Subject identification

time

Times of covariate observations

covnames...

Columns with each covariate observations in the dataset for each subject and time

parnames...

Columns with each parameter in the model and the icen summary for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian parameter means and medians

icen

The type of summarized Bayesian posterior individual parameter values: mean or median.

Details

For each subject, makeCov extracts covariate information and Bayesian posterior parameter estimates. This output of this function is suitable for exploration of covariate-parameter, covariate-time, or parameter-time relationships.

See also

Author

Michael Neely

Examples

cov <- makeCov(NPex$NPdata)
cov
#>    id time   wt africa age gender height       Ka        Ke        V     Tlag1
#> 1   1    0 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 2   1   24 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 3   1   48 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 4   1   72 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 5   1   96 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 6   1  120 46.7      1  21      1    160 0.440192 0.0246167  66.3882 0.5544030
#> 7   2    0 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 8   2   24 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 9   2   48 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 10  2   72 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 11  2   96 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 12  2  120 66.5      1  30      1    174 0.738613 0.0398178 119.4690 0.0259495
#> 13  3    0 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 14  3   24 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 15  3   48 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 16  3   72 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 17  3   96 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 18  3  120 46.7      1  24      0    164 0.899944 0.0431027 108.6490 2.0959200
#> 19  4    0 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 20  4   24 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 21  4   48 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 22  4   72 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 23  4   96 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 24  4  120 50.8      1  25      1    165 0.897550 0.0564307 119.8190 0.6883040
#> 25  5    0 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 26  5   24 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 27  5   48 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 28  5   72 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 29  5   96 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 30  5  120 65.8      1  22      1    181 0.105273 0.0675058 113.3430 0.0186591
#> 31  6    0 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 32  6   24 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 33  6   48 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 34  6   72 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 35  6   96 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 36  6  120 65.0      1  23      1    177 0.895237 0.0348831  71.8617 1.9978800
#> 37  7    0 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 38  7   24 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 39  7   48 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 40  7   72 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 41  7   96 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 42  7  120 51.7      1  27      0    161 0.215198 0.0832836  35.2243 1.7966900
#> 43  8    0 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 44  8   24 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 45  8   48 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 46  8   72 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 47  8   96 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 48  8  120 51.2      1  22      1    163 0.895494 0.0348882  71.8466 1.9985200
#> 49  9    0 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 50  9   24 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 51  9   48 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 52  9   72 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 53  9   96 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 54  9  120 55.0      1  23      1    174 0.790042 0.0439417 101.7520 0.8801720
#> 55 10    0 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 56 10   24 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 57 10   48 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 58 10   72 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 59 10   96 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 60 10  120 52.1      1  32      1    163 0.655770 0.0615898  61.6882 0.8013670
#> 61 11    0 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 62 11   24 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 63 11   48 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 64 11   72 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 65 11   96 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 66 11  120 56.5      1  34      1    165 0.583214 0.0683228  73.1084 1.3385500
#> 67 12    0 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 68 12   24 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 69 12   48 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 70 12   72 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 71 12   96 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 72 12  120 47.9      1  54      0    160 0.470094 0.0306855  91.8654 1.0252500
#> 73 13    0 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 74 13   24 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 75 13   48 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 76 13   72 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 77 13   96 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 78 13  120 60.5      1  24      1    180 0.215198 0.0832837  35.2243 1.7957700
#> 79 14    0 59.2      1  26      1    174 0.555242 0.0439039 117.8380 0.2868650
#> 80 14   24 59.2      1  26      1    174 0.555242 0.0439039 117.8380 0.2868650
#> 81 14   48 59.2      1  26      1    174 0.555242 0.0439039 117.8380 0.2868650
#> 82 14   72 59.2      1  26      1    174 0.555242 0.0439039 117.8380 0.2868650
#> 83 14   96 59.2      1  26      1    174 0.555242 0.0439039 117.8380 0.2868650
#>    icen
#> 1  mean
#> 2  mean
#> 3  mean
#> 4  mean
#> 5  mean
#> 6  mean
#> 7  mean
#> 8  mean
#> 9  mean
#> 10 mean
#> 11 mean
#> 12 mean
#> 13 mean
#> 14 mean
#> 15 mean
#> 16 mean
#> 17 mean
#> 18 mean
#> 19 mean
#> 20 mean
#> 21 mean
#> 22 mean
#> 23 mean
#> 24 mean
#> 25 mean
#> 26 mean
#> 27 mean
#> 28 mean
#> 29 mean
#> 30 mean
#> 31 mean
#> 32 mean
#> 33 mean
#> 34 mean
#> 35 mean
#> 36 mean
#> 37 mean
#> 38 mean
#> 39 mean
#> 40 mean
#> 41 mean
#> 42 mean
#> 43 mean
#> 44 mean
#> 45 mean
#> 46 mean
#> 47 mean
#> 48 mean
#> 49 mean
#> 50 mean
#> 51 mean
#> 52 mean
#> 53 mean
#> 54 mean
#> 55 mean
#> 56 mean
#> 57 mean
#> 58 mean
#> 59 mean
#> 60 mean
#> 61 mean
#> 62 mean
#> 63 mean
#> 64 mean
#> 65 mean
#> 66 mean
#> 67 mean
#> 68 mean
#> 69 mean
#> 70 mean
#> 71 mean
#> 72 mean
#> 73 mean
#> 74 mean
#> 75 mean
#> 76 mean
#> 77 mean
#> 78 mean
#> 79 mean
#> 80 mean
#> 81 mean
#> 82 mean
#> 83 mean
#>  [ reached 'max' / getOption("max.print") -- omitted 157 rows ]
names(cov)
#>  [1] "id"     "time"   "wt"     "africa" "age"    "gender" "height" "Ka"    
#>  [9] "Ke"     "V"      "Tlag1"  "icen"