Extract covariate data

Generates an data.frame with subject-specific covariate data from an NPAG or IT2B object

Usage

makeCov(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeCov is a dataframe of class PMcov, which has the following columns:

id

Subject identification

time

Times of covariate observations

covnames...

Columns with each covariate observations in the dataset for each subject and time

parnames...

Columns with each parameter in the model and the icen summary for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian parameter means and medians

icen

The type of summarized Bayesian posterior individual parameter values: mean or median.

Details

For each subject, makeCov extracts covariate information and Bayesian posterior parameter estimates. This output of this function is suitable for exploration of covariate-parameter, covariate-time, or parameter-time relationships.

See also

Author

Michael Neely

Examples

library(PmetricsData)
cov <- makeCov(NPex$NPdata)
cov
#>     id time   wt africa age gender height block        Ka         Ke         V
#> 1    1    0 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 2    1   24 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 3    1   48 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 4    1   72 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 5    1   96 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 6    1  120 46.7      1  21      1    160     1 0.4401920 0.02461670  66.38820
#> 7    2    0 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 8    2   24 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 9    2   48 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 10   2   72 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 11   2   96 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 12   2  120 66.5      1  30      1    174     1 0.7386130 0.03981780 119.46900
#> 13   3    0 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 14   3   24 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 15   3   48 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 16   3   72 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 17   3   96 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 18   3  120 46.7      1  24      0    164     1 0.8999440 0.04310270 108.64900
#> 19   4    0 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 20   4   24 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 21   4   48 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 22   4   72 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 23   4   96 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 24   4  120 50.8      1  25      1    165     1 0.8975500 0.05643070 119.81900
#> 25   5    0 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 26   5   24 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 27   5   48 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 28   5   72 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 29   5   96 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 30   5  120 65.8      1  22      1    181     1 0.1052730 0.06750580 113.34300
#> 31   6    0 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 32   6   24 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 33   6   48 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 34   6   72 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 35   6   96 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 36   6  120 65.0      1  23      1    177     1 0.8952370 0.03488310  71.86170
#> 37   7    0 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 38   7   24 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 39   7   48 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 40   7   72 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 41   7   96 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 42   7  120 51.7      1  27      0    161     1 0.2151980 0.08328360  35.22430
#> 43   8    0 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 44   8   24 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 45   8   48 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 46   8   72 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 47   8   96 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 48   8  120 51.2      1  22      1    163     1 0.8954940 0.03488820  71.84660
#> 49   9    0 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 50   9   24 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 51   9   48 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 52   9   72 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 53   9   96 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 54   9  120 55.0      1  23      1    174     1 0.7900420 0.04394170 101.75200
#> 55  10    0 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 56  10   24 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 57  10   48 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 58  10   72 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 59  10   96 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 60  10  120 52.1      1  32      1    163     1 0.6557700 0.06158980  61.68820
#> 61  11    0 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 62  11   24 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 63  11   48 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 64  11   72 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 65  11   96 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 66  11  120 56.5      1  34      1    165     1 0.5832140 0.06832280  73.10840
#> 67  12    0 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 68  12   24 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 69  12   48 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 70  12   72 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 71  12   96 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 72  12  120 47.9      1  54      0    160     1 0.4700940 0.03068550  91.86540
#> 73  13    0 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 74  13   24 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 75  13   48 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 76  13   72 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 77  13   96 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 78  13  120 60.5      1  24      1    180     1 0.2151980 0.08328370  35.22430
#> 79  14    0 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 80  14   24 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 81  14   48 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 82  14   72 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 83  14   96 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 84  14  120 59.2      1  26      1    174     1 0.5552420 0.04390390 117.83800
#> 85  15    0 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 86  15   24 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 87  15   48 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 88  15   72 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 89  15   96 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 90  15  120 43.0      1  19      0    150     1 0.7957240 0.03421020  72.19660
#> 91  16    0 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 92  16   24 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 93  16   48 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 94  16   72 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 95  16   96 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 96  16  120 64.4      1  25      1    173     1 0.7529810 0.03529900  89.67020
#> 97  17    0 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 98  17   24 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 99  17   48 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 100 17   72 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 101 17   96 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 102 17  120 54.8      1  23      1    170     1 0.8913430 0.07341410  63.31680
#> 103 18    0 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 104 18   24 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 105 18   48 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 106 18   72 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 107 18   96 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 108 18  120 44.3      1  20      0    164     1 0.8946330 0.02319190  75.92770
#> 109 19    0 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 110 19   24 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 111 19   48 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 112 19   72 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 113 19   96 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 114 19  120 50.0      1  36      1    168     1 0.6575970 0.06211690  30.98520
#> 115 20    0 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 116 20   24 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 117 20   48 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 118 20   72 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 119 20   96 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 120 20  120 59.0      1  31      1    170     1 0.2151980 0.08328370  35.22430
#> 121  1    0 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 122  1   24 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 123  1   48 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 124  1   72 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 125  1   96 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 126  1  120 46.7      1  21      1    160     1 0.4320679 0.02425660  66.45764
#> 127  2    0 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 128  2   24 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 129  2   48 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 130  2   72 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 131  2   96 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 132  2  120 66.5      1  30      1    174     1 0.7438061 0.03914022 119.52708
#> 133  3    0 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 134  3   24 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 135  3   48 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 136  3   72 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 137  3   96 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 138  3  120 46.7      1  24      0    164     1 0.8960000 0.04307500 108.75000
#> 139  4    0 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 140  4   24 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 141  4   48 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 142  4   72 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 143  4   96 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 144  4  120 50.8      1  25      1    165     1 0.8959996 0.05594505 119.54995
#> 145  5    0 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 146  5   24 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 147  5   48 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 148  5   72 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 149  5   96 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 150  5  120 65.8      1  22      1    181     1 0.1040173 0.06782285 113.25195
#> 151  6    0 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 152  6   24 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 153  6   48 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 154  6   72 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 155  6   96 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 156  6  120 65.0      1  23      1    177     1 0.8959943 0.03515430  71.85063
#> 157  7    0 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 158  7   24 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 159  7   48 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 160  7   72 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 161  7   96 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 162  7  120 51.7      1  27      0    161     1 0.2160000 0.08366500  34.95000
#> 163  8    0 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 164  8   24 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 165  8   48 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 166  8   72 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 167  8   96 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 168  8  120 51.2      1  22      1    163     1 0.8959968 0.03515465  71.85031
#> 169  9    0 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 170  9   24 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 171  9   48 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 172  9   72 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 173  9   96 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 174  9  120 55.0      1  23      1    174     1 0.7919535 0.04406497 101.55520
#> 175 10    0 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 176 10   24 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 177 10   48 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 178 10   72 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 179 10   96 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 180 10  120 52.1      1  32      1    163     1 0.6560165 0.06188295  61.95186
#> 181 11    0 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 182 11   24 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 183 11   48 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 184 11   72 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 185 11   96 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 186 11  120 56.5      1  34      1    165     1 0.5840040 0.06782547  72.74957
#> 187 12    0 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 188 12   24 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 189 12   48 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 190 12   72 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 191 12   96 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 192 12  120 47.9      1  54      0    160     1 0.4641139 0.03021910  91.63719
#> 193 13    0 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 194 13   24 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 195 13   48 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 196 13   72 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 197 13   96 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 198 13  120 60.5      1  24      1    180     1 0.2160000 0.08366500  34.95000
#> 199 14    0 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 200 14   24 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 201 14   48 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 202 14   72 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 203 14   96 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 204 14  120 59.2      1  26      1    174     1 0.5520477 0.04406093 117.75206
#> 205 15    0 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 206 15   24 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 207 15   48 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 208 15   72 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 209 15   96 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 210 15  120 43.0      1  19      0    150     1 0.7999679 0.03416107  71.85822
#> 211 16    0 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 212 16   24 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 213 16   48 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 214 16   72 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 215 16   96 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 216 16  120 64.4      1  25      1    173     1 0.7519847 0.03515311  89.85172
#> 217 17    0 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 218 17   24 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 219 17   48 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 220 17   72 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 221 17   96 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 222 17  120 54.8      1  23      1    170     1 0.8959356 0.07375703  62.85724
#> 223 18    0 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 224 18   24 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 225 18   48 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 226 18   72 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 227 18   96 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 228 18  120 44.3      1  20      0    164     1 0.8959946 0.02327567  76.34939
#> 229 19    0 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 230 19   24 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 231 19   48 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 232 19   72 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 233 19   96 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 234 19  120 50.0      1  36      1    168     1 0.6560000 0.06188500  31.35000
#> 235 20    0 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#> 236 20   24 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#> 237 20   48 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#> 238 20   72 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#> 239 20   96 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#> 240 20  120 59.0      1  31      1    170     1 0.2160000 0.08366500  34.95000
#>            lag   icen
#> 1   0.55440300   mean
#> 2   0.55440300   mean
#> 3   0.55440300   mean
#> 4   0.55440300   mean
#> 5   0.55440300   mean
#> 6   0.55440300   mean
#> 7   0.02594950   mean
#> 8   0.02594950   mean
#> 9   0.02594950   mean
#> 10  0.02594950   mean
#> 11  0.02594950   mean
#> 12  0.02594950   mean
#> 13  2.09592000   mean
#> 14  2.09592000   mean
#> 15  2.09592000   mean
#> 16  2.09592000   mean
#> 17  2.09592000   mean
#> 18  2.09592000   mean
#> 19  0.68830400   mean
#> 20  0.68830400   mean
#> 21  0.68830400   mean
#> 22  0.68830400   mean
#> 23  0.68830400   mean
#> 24  0.68830400   mean
#> 25  0.01865910   mean
#> 26  0.01865910   mean
#> 27  0.01865910   mean
#> 28  0.01865910   mean
#> 29  0.01865910   mean
#> 30  0.01865910   mean
#> 31  1.99788000   mean
#> 32  1.99788000   mean
#> 33  1.99788000   mean
#> 34  1.99788000   mean
#> 35  1.99788000   mean
#> 36  1.99788000   mean
#> 37  1.79669000   mean
#> 38  1.79669000   mean
#> 39  1.79669000   mean
#> 40  1.79669000   mean
#> 41  1.79669000   mean
#> 42  1.79669000   mean
#> 43  1.99852000   mean
#> 44  1.99852000   mean
#> 45  1.99852000   mean
#> 46  1.99852000   mean
#> 47  1.99852000   mean
#> 48  1.99852000   mean
#> 49  0.88017200   mean
#> 50  0.88017200   mean
#> 51  0.88017200   mean
#> 52  0.88017200   mean
#> 53  0.88017200   mean
#> 54  0.88017200   mean
#> 55  0.80136700   mean
#> 56  0.80136700   mean
#> 57  0.80136700   mean
#> 58  0.80136700   mean
#> 59  0.80136700   mean
#> 60  0.80136700   mean
#> 61  1.33855000   mean
#> 62  1.33855000   mean
#> 63  1.33855000   mean
#> 64  1.33855000   mean
#> 65  1.33855000   mean
#> 66  1.33855000   mean
#> 67  1.02525000   mean
#> 68  1.02525000   mean
#> 69  1.02525000   mean
#> 70  1.02525000   mean
#> 71  1.02525000   mean
#> 72  1.02525000   mean
#> 73  1.79577000   mean
#> 74  1.79577000   mean
#> 75  1.79577000   mean
#> 76  1.79577000   mean
#> 77  1.79577000   mean
#> 78  1.79577000   mean
#> 79  0.28686500   mean
#> 80  0.28686500   mean
#> 81  0.28686500   mean
#> 82  0.28686500   mean
#> 83  0.28686500   mean
#> 84  0.28686500   mean
#> 85  1.05844000   mean
#> 86  1.05844000   mean
#> 87  1.05844000   mean
#> 88  1.05844000   mean
#> 89  1.05844000   mean
#> 90  1.05844000   mean
#> 91  0.68794600   mean
#> 92  0.68794600   mean
#> 93  0.68794600   mean
#> 94  0.68794600   mean
#> 95  0.68794600   mean
#> 96  0.68794600   mean
#> 97  1.17989000   mean
#> 98  1.17989000   mean
#> 99  1.17989000   mean
#> 100 1.17989000   mean
#> 101 1.17989000   mean
#> 102 1.17989000   mean
#> 103 1.76041000   mean
#> 104 1.76041000   mean
#> 105 1.76041000   mean
#> 106 1.76041000   mean
#> 107 1.76041000   mean
#> 108 1.76041000   mean
#> 109 1.92418000   mean
#> 110 1.92418000   mean
#> 111 1.92418000   mean
#> 112 1.92418000   mean
#> 113 1.92418000   mean
#> 114 1.92418000   mean
#> 115 1.79601000   mean
#> 116 1.79601000   mean
#> 117 1.79601000   mean
#> 118 1.79601000   mean
#> 119 1.79601000   mean
#> 120 1.79601000   mean
#> 121 0.54033955 median
#> 122 0.54033955 median
#> 123 0.54033955 median
#> 124 0.54033955 median
#> 125 0.54033955 median
#> 126 0.54033955 median
#> 127 0.02101887 median
#> 128 0.02101887 median
#> 129 0.02101887 median
#> 130 0.02101887 median
#> 131 0.02101887 median
#> 132 0.02101887 median
#> 133 2.10000000 median
#> 134 2.10000000 median
#> 135 2.10000000 median
#> 136 2.10000000 median
#> 137 2.10000000 median
#> 138 2.10000000 median
#> 139 0.69999785 median
#> 140 0.69999785 median
#> 141 0.69999785 median
#> 142 0.69999785 median
#> 143 0.69999785 median
#> 144 0.69999785 median
#> 145 0.02008672 median
#> 146 0.02008672 median
#> 147 0.02008672 median
#> 148 0.02008672 median
#> 149 0.02008672 median
#> 150 0.02008672 median
#> 151 1.97997170 median
#> 152 1.97997170 median
#> 153 1.97997170 median
#> 154 1.97997170 median
#> 155 1.97997170 median
#> 156 1.97997170 median
#> 157 1.81206790 median
#> 158 1.81206790 median
#> 159 1.81206790 median
#> 160 1.81206790 median
#> 161 1.81206790 median
#> 162 1.81206790 median
#> 163 1.97998410 median
#> 164 1.97998410 median
#> 165 1.97998410 median
#> 166 1.97998410 median
#> 167 1.97998410 median
#> 168 1.97998410 median
#> 169 0.89976788 median
#> 170 0.89976788 median
#> 171 0.89976788 median
#> 172 0.89976788 median
#> 173 0.89976788 median
#> 174 0.89976788 median
#> 175 0.82008218 median
#> 176 0.82008218 median
#> 177 0.82008218 median
#> 178 0.82008218 median
#> 179 0.82008218 median
#> 180 0.82008218 median
#> 181 1.33998020 median
#> 182 1.33998020 median
#> 183 1.33998020 median
#> 184 1.33998020 median
#> 185 1.33998020 median
#> 186 1.33998020 median
#> 187 1.01977840 median
#> 188 1.01977840 median
#> 189 1.01977840 median
#> 190 1.01977840 median
#> 191 1.01977840 median
#> 192 1.01977840 median
#> 193 1.81054350 median
#> 194 1.81054350 median
#> 195 1.81054350 median
#> 196 1.81054350 median
#> 197 1.81054350 median
#> 198 1.81054350 median
#> 199 0.29990845 median
#> 200 0.29990845 median
#> 201 0.29990845 median
#> 202 0.29990845 median
#> 203 0.29990845 median
#> 204 0.29990845 median
#> 205 1.05983960 median
#> 206 1.05983960 median
#> 207 1.05983960 median
#> 208 1.05983960 median
#> 209 1.05983960 median
#> 210 1.05983960 median
#> 211 0.70007653 median
#> 212 0.70007653 median
#> 213 0.70007653 median
#> 214 0.70007653 median
#> 215 0.70007653 median
#> 216 0.70007653 median
#> 217 1.18032190 median
#> 218 1.18032190 median
#> 219 1.18032190 median
#> 220 1.18032190 median
#> 221 1.18032190 median
#> 222 1.18032190 median
#> 223 1.77997280 median
#> 224 1.77997280 median
#> 225 1.77997280 median
#> 226 1.77997280 median
#> 227 1.77997280 median
#> 228 1.77997280 median
#> 229 1.94000010 median
#> 230 1.94000010 median
#> 231 1.94000010 median
#> 232 1.94000010 median
#> 233 1.94000010 median
#> 234 1.94000010 median
#> 235 1.81096390 median
#> 236 1.81096390 median
#> 237 1.81096390 median
#> 238 1.81096390 median
#> 239 1.81096390 median
#> 240 1.81096390 median
names(cov)
#>  [1] "id"     "time"   "wt"     "africa" "age"    "gender" "height" "block" 
#>  [9] "Ka"     "Ke"     "V"      "lag"    "icen"