Extract covariate data

Generates an data.frame with subject-specific covariate data from an NPAG or IT2B object

Usage

makeCov(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeCov is a dataframe of class PMcov, which has the following columns:

id

Subject identification

time

Times of covariate observations

covnames...

Columns with each covariate observations in the dataset for each subject and time

parnames...

Columns with each parameter in the model and the icen summary for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian parameter means and medians

icen

The type of summarized Bayesian posterior individual parameter values: mean or median.

Details

For each subject, makeCov extracts covariate information and Bayesian posterior parameter estimates. This output of this function is suitable for exploration of covariate-parameter, covariate-time, or parameter-time relationships.

See also

Author

Michael Neely

Examples

data(NPdata.1)
#> Warning: data set ‘NPdata.1’ not found
cov <- makeCov(NPdata.1)
#> Error in makeCov(NPdata.1): object 'NPdata.1' not found
cov
#> function (x, y = NULL, use = "everything", method = c("pearson", 
#>     "kendall", "spearman")) 
#> {
#>     na.method <- pmatch(use, c("all.obs", "complete.obs", "pairwise.complete.obs", 
#>         "everything", "na.or.complete"))
#>     if (is.na(na.method)) 
#>         stop("invalid 'use' argument")
#>     method <- match.arg(method)
#>     if (is.data.frame(y)) 
#>         y <- as.matrix(y)
#>     if (is.data.frame(x)) 
#>         x <- as.matrix(x)
#>     if (!is.matrix(x) && is.null(y)) 
#>         stop("supply both 'x' and 'y' or a matrix-like 'x'")
#>     stopifnot(is.numeric(x) || is.logical(x), is.atomic(x))
#>     if (!is.null(y)) 
#>         stopifnot(is.numeric(y) || is.logical(y), is.atomic(y))
#>     Rank <- function(u) {
#>         if (length(u) == 0L) 
#>             u
#>         else if (is.matrix(u)) {
#>             if (nrow(u) > 1L) 
#>                 apply(u, 2L, rank, na.last = "keep")
#>             else row(u)
#>         }
#>         else rank(u, na.last = "keep")
#>     }
#>     if (method == "pearson") 
#>         .Call(C_cov, x, y, na.method, method == "kendall")
#>     else if (na.method %in% c(2L, 5L)) {
#>         if (is.null(y)) {
#>             .Call(C_cov, Rank(na.omit(x)), NULL, na.method, method == 
#>                 "kendall")
#>         }
#>         else {
#>             nas <- attr(na.omit(cbind(x, y)), "na.action")
#>             dropNA <- function(x, nas) {
#>                 if (length(nas)) {
#>                   if (is.matrix(x)) 
#>                     x[-nas, , drop = FALSE]
#>                   else x[-nas]
#>                 }
#>                 else x
#>             }
#>             .Call(C_cov, Rank(dropNA(x, nas)), Rank(dropNA(y, 
#>                 nas)), na.method, method == "kendall")
#>         }
#>     }
#>     else if (na.method != 3L) {
#>         x <- Rank(x)
#>         if (!is.null(y)) 
#>             y <- Rank(y)
#>         .Call(C_cov, x, y, na.method, method == "kendall")
#>     }
#>     else stop("cannot handle 'pairwise.complete.obs'")
#> }
#> <bytecode: 0x10a3c03d0>
#> <environment: namespace:stats>
names(cov)
#> NULL