Summarize Pmetrics Run Cycle Information

Parses the cycle information from an NPAG or an IT2B object.

Usage

makeCycle(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeCycle is a list of class PMcycle, which has 8 objects from NPAG or 6 objects from IT2B :

names

Vector of names of the random parameters

#'

names

Vector of names of the random parameters

cycnum

Vector cycle numbers, which may start at numbers greater than 1 if a non-uniform prior was specified for the run (NPAG only)

ll

Matrix of cycle number and -2*Log-likelihood at each cycle

gamlam

A matrix of cycle number and gamma or lambda at each cycle

mean

A matrix of cycle number and the mean of each random parameter at each cycle, normalized to initial mean

sd

A matrix of cycle number and the standard deviation of each random parameter at each cycle, normalized to initial standard deviation

median

A matrix of cycle number and the median of each random parameter at each cycle, normalized to initial median

aic

A matrix of cycle number and Akaike Information Criterion at each cycle

bic

A matrix of cycle number and Bayesian (Schwartz) Information Criterion at each cycle

A plot method exists in plot for PMcycle objects.

Details

This function will parse the output of NPparse or ITparse to generate a list suitable for analysis and plotting of NPAG or IT2B cycle information.

See also

Author

Michael Neely

Examples

cycle <- makeCycle(NPex$NPdata)
cycle
#> $names
#> [1] "Ka"  "Ke"  "V"   "lag"
#> 
#> $cycnum
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100
#> 
#> $ll
#>   [1] 507.3346 481.2885 460.5295 456.6656 456.4747 456.4747 454.6416 453.0673
#>   [9] 452.7149 452.6453 452.6453 448.5969 448.4804 448.3917 448.3405 448.3009
#>  [17] 448.2713 448.2713 446.7067 445.8884 445.5819 445.5588 445.5471 445.5426
#>  [25] 445.5426 444.4183 443.9666 443.4177 443.0373 442.8305 442.7360 442.7092
#>  [33] 442.6847 442.6652 442.5090 442.2605 442.1442 442.1415 442.1399 442.1396
#>  [41] 442.1369 442.1322 442.1288 442.1267 442.1259 442.1214 442.1164 442.1128
#>  [49] 442.1106 442.1077 442.1014 442.0966 442.0932 442.0902 442.0829 442.0771
#>  [57] 442.0728 442.0681 442.0600 442.0535 442.0486 442.0447 442.0353 442.0278
#>  [65] 442.0218 442.0172 442.0141 442.0124 442.0120 442.0120 441.6645 441.4615
#>  [73] 441.3766 441.3248 441.2395 441.1680 441.1068 441.0444 440.9916 440.8840
#>  [81] 440.7468 440.6802 440.5175 440.4220 440.4022 440.3955 440.3622 440.3499
#>  [89] 440.3384 440.3066 440.2895 440.2692 440.2470 440.2440 440.2243 440.2052
#>  [97] 440.2048 440.1825 440.1657 440.1657
#> 
#> $gamlam
#>            [,1]
#>   [1,] 4.545455
#>   [2,] 3.787879
#>   [3,] 2.705628
#>   [4,] 2.705628
#>   [5,] 2.705628
#>   [6,] 2.705628
#>   [7,] 2.705628
#>   [8,] 2.576788
#>   [9,] 2.576788
#>  [10,] 2.576788
#>  [11,] 2.576788
#>  [12,] 2.544976
#>  [13,] 2.544976
#>  [14,] 2.513557
#>  [15,] 2.513557
#>  [16,] 2.513557
#>  [17,] 2.513557
#>  [18,] 2.513557
#>  [19,] 2.488670
#>  [20,] 2.488670
#>  [21,] 2.464030
#>  [22,] 2.464030
#>  [23,] 2.464030
#>  [24,] 2.464030
#>  [25,] 2.464030
#>  [26,] 2.464030
#>  [27,] 2.464030
#>  [28,] 2.439633
#>  [29,] 2.439633
#>  [30,] 2.439633
#>  [31,] 2.439633
#>  [32,] 2.439633
#>  [33,] 2.439633
#>  [34,] 2.439633
#>  [35,] 2.439633
#>  [36,] 2.439633
#>  [37,] 2.439633
#>  [38,] 2.439633
#>  [39,] 2.439633
#>  [40,] 2.439633
#>  [41,] 2.439633
#>  [42,] 2.439633
#>  [43,] 2.439633
#>  [44,] 2.439633
#>  [45,] 2.439633
#>  [46,] 2.439633
#>  [47,] 2.439633
#>  [48,] 2.439633
#>  [49,] 2.439633
#>  [50,] 2.439633
#>  [51,] 2.439633
#>  [52,] 2.439633
#>  [53,] 2.439633
#>  [54,] 2.439633
#>  [55,] 2.439633
#>  [56,] 2.439633
#>  [57,] 2.439633
#>  [58,] 2.439633
#>  [59,] 2.439633
#>  [60,] 2.439633
#>  [61,] 2.439633
#>  [62,] 2.439633
#>  [63,] 2.439633
#>  [64,] 2.439633
#>  [65,] 2.439633
#>  [66,] 2.439633
#>  [67,] 2.439633
#>  [68,] 2.439633
#>  [69,] 2.439633
#>  [70,] 2.439633
#>  [71,] 2.439633
#>  [72,] 2.439633
#>  [73,] 2.415478
#>  [74,] 2.415478
#>  [75,] 2.415478
#>  [76,] 2.415478
#>  [77,] 2.415478
#>  [78,] 2.415478
#>  [79,] 2.415478
#>  [80,] 2.415478
#>  [81,] 2.415478
#>  [82,] 2.415478
#>  [83,] 2.415478
#>  [84,] 2.415478
#>  [85,] 2.415478
#>  [86,] 2.415478
#>  [87,] 2.415478
#>  [88,] 2.415478
#>  [89,] 2.415478
#>  [90,] 2.415478
#>  [91,] 2.415478
#>  [92,] 2.415478
#>  [93,] 2.415478
#>  [94,] 2.415478
#>  [95,] 2.415478
#>  [96,] 2.415478
#>  [97,] 2.415478
#>  [98,] 2.415478
#>  [99,] 2.415478
#> [100,] 2.415478
#> 
#> $mean
#>      par
#> cycle        Ka        Ke         V       lag
#>   1   1.0000000 1.0000000 1.0000000 1.0000000
#>   2   1.0149965 1.0036643 0.9925630 0.9965626
#>   3   1.0313964 1.0060456 0.9889442 0.9966983
#>   4   1.0195397 1.0061282 0.9953752 0.9840417
#>   5   0.9937913 1.0060956 0.9946988 0.9833331
#>   6   0.9937913 1.0060956 0.9946988 0.9833331
#>   7   1.0052473 1.0058939 1.0057208 0.9541829
#>   8   1.0047262 1.0055674 1.0043512 0.9489062
#>   9   1.0105375 1.0053677 1.0025242 0.9499842
#>   10  1.0171156 1.0052275 1.0026923 0.9500520
#>   11  1.0171156 1.0052275 1.0026923 0.9500520
#>   12  1.0214872 1.0010543 1.0001642 0.9398529
#>   13  1.0267543 1.0010428 1.0003115 0.9323448
#>   14  1.0333324 1.0010524 1.0003388 0.9414585
#>   15  1.0367525 1.0010524 1.0003427 0.9414962
#>   16  1.0401793 1.0010505 1.0003479 0.9415339
#>   17  1.0469074 1.0010447 1.0003662 0.9417072
#>   18  1.0469074 1.0010447 1.0003662 0.9417072
#>   19  1.0431723 0.9991454 0.9988233 0.9348173
#>   20  1.0507745 0.9927848 1.0026493 0.9336942
#>   21  1.0510152 0.9915787 1.0039029 0.9343500
#>   22  1.0508685 0.9915999 1.0039120 0.9343726
#>   23  1.0523213 0.9916172 1.0039211 0.9344027
#>   24  1.0537675 0.9916325 1.0039303 0.9344329
#>   25  1.0537675 0.9916325 1.0039303 0.9344329
#>   26  1.0528177 0.9953601 1.0038000 0.9304075
#>   27  1.0520706 0.9939121 1.0027926 0.9371316
#>   28  1.0509658 0.9940830 1.0039420 0.9364908
#>   29  1.0522586 0.9940600 1.0052139 0.9362647
#>   30  1.0524549 0.9928021 1.0059319 0.9344857
#>   31  1.0525588 0.9928194 1.0066577 0.9343801
#>   32  1.0531409 0.9927886 1.0067164 0.9325634
#>   33  1.0539984 0.9927963 1.0067073 0.9308146
#>   34  1.0532794 0.9928021 1.0072924 0.9308749
#>   35  1.0519222 0.9915269 1.0078944 0.9311312
#>   36  1.0513896 0.9915845 1.0083205 0.9309050
#>   37  1.0515396 0.9916114 1.0088535 0.9308523
#>   38  1.0507168 0.9916114 1.0088522 0.9308221
#>   39  1.0498939 0.9916133 1.0088509 0.9307995
#>   40  1.0490710 0.9916152 1.0088496 0.9307693
#>   41  1.0490644 0.9916095 1.0088639 0.9290054
#>   42  1.0482415 0.9916095 1.0088613 0.9289828
#>   43  1.0474203 0.9916114 1.0088600 0.9289602
#>   44  1.0465991 0.9916133 1.0088574 0.9289300
#>   45  1.0457778 0.9916133 1.0088548 0.9288999
#>   46  1.0457630 0.9916095 1.0088692 0.9271284
#>   47  1.0449434 0.9916095 1.0088666 0.9270983
#>   48  1.0441238 0.9916114 1.0088639 0.9270756
#>   49  1.0433059 0.9916133 1.0088613 0.9270530
#>   50  1.0432845 0.9916075 1.0088757 0.9252589
#>   51  1.0424665 0.9916095 1.0088731 0.9252363
#>   52  1.0416486 0.9916114 1.0088692 0.9252062
#>   53  1.0408323 0.9916133 1.0088653 0.9251836
#>   54  1.0408059 0.9916075 1.0088796 0.9233744
#>   55  1.0399896 0.9916095 1.0088757 0.9233518
#>   56  1.0391750 0.9916114 1.0088718 0.9233367
#>   57  1.0383620 0.9916114 1.0088679 0.9233065
#>   58  1.0383307 0.9916075 1.0088822 0.9214898
#>   59  1.0375177 0.9916095 1.0088770 0.9214672
#>   60  1.0367064 0.9916114 1.0088718 0.9214522
#>   61  1.0358950 0.9916133 1.0088679 0.9214295
#>   62  1.0358637 0.9916075 1.0088809 0.9195978
#>   63  1.0360715 0.9916210 1.0094712 0.9196807
#>   64  1.0352684 0.9916248 1.0094699 0.9196656
#>   65  1.0344670 0.9916268 1.0094673 0.9196430
#>   66  1.0336655 0.9916287 1.0094660 0.9196204
#>   67  1.0328657 0.9916306 1.0094634 0.9195978
#>   68  1.0320659 0.9916325 1.0094621 0.9195827
#>   69  1.0312661 0.9916344 1.0094608 0.9195601
#>   70  1.0312661 0.9916344 1.0094608 0.9195601
#>   71  1.0312299 0.9904361 1.0096875 0.9180977
#>   72  1.0310402 0.9904188 1.0103586 0.9164167
#>   73  1.0317031 0.9898426 1.0106023 0.9163036
#>   74  1.0325392 0.9898484 1.0105997 0.9144341
#>   75  1.0329943 0.9898522 1.0105841 0.9125420
#>   76  1.0339112 0.9898292 1.0106075 0.9134315
#>   77  1.0346335 0.9904015 1.0109894 0.9106424
#>   78  1.0350623 0.9898234 1.0106167 0.9087654
#>   79  1.0340448 0.9891686 1.0112513 0.9076874
#>   80  1.0338980 0.9897850 1.0118664 0.9058104
#>   81  1.0336869 0.9892204 1.0116579 0.9058330
#>   82  1.0338304 0.9897869 1.0116970 0.9057275
#>   83  1.0342212 0.9904284 1.0113295 0.9067075
#>   84  1.0345939 0.9898350 1.0106062 0.9067225
#>   85  1.0349979 0.9898330 1.0109724 0.9048380
#>   86  1.0354102 0.9892415 1.0109711 0.9029535
#>   87  1.0358505 0.9892473 1.0113243 0.9011066
#>   88  1.0359016 0.9892531 1.0116696 0.8992598
#>   89  1.0359577 0.9886692 1.0116149 0.8975109
#>   90  1.0359000 0.9886615 1.0113569 0.8972471
#>   91  1.0359379 0.9886673 1.0116488 0.8955359
#>   92  1.0359956 0.9880835 1.0116266 0.8937041
#>   93  1.0359923 0.9880835 1.0119902 0.8936966
#>   94  1.0359923 0.9880835 1.0123551 0.8936966
#>   95  1.0359890 0.9874920 1.0123551 0.8936966
#>   96  1.0359874 0.9874920 1.0127199 0.8936966
#>   97  1.0359874 0.9874920 1.0130848 0.8936966
#>   98  1.0359841 0.9869024 1.0130848 0.8936966
#>   99  1.0359824 0.9869024 1.0134497 0.8936966
#>   100 1.0359824 0.9869024 1.0134497 0.8936966
#> 
#> $sd
#>      par
#> cycle       Ka        Ke         V      lag
#>   1   1.000000 1.0000000 1.0000000 1.000000
#>   2   1.032320 1.0049040 0.9819494 1.002131
#>   3   1.066454 1.0098532 0.9714715 1.005992
#>   4   1.044000 1.0101540 0.9731300 1.181892
#>   5   1.032260 1.0101791 0.9729111 1.181677
#>   6   1.032260 1.0101791 0.9729111 1.181677
#>   7   1.028150 1.0113274 1.0083445 1.098919
#>   8   1.028072 1.0118940 1.0069568 1.136862
#>   9   1.051739 1.0120043 0.9996920 1.137023
#>   10  1.058213 1.0120093 0.9997440 1.136908
#>   11  1.058213 1.0120093 0.9997440 1.136908
#>   12  1.080215 0.9968711 1.0140064 1.128016
#>   13  1.085064 0.9968761 1.0145073 1.103048
#>   14  1.095502 0.9969162 1.0143589 1.127994
#>   15  1.101306 0.9969212 1.0143589 1.127982
#>   16  1.108391 0.9969262 1.0143551 1.127971
#>   17  1.111452 0.9969463 1.0143440 1.127924
#>   18  1.111452 0.9969463 1.0143440 1.127924
#>   19  1.102235 0.9993832 1.0111235 1.095480
#>   20  1.092846 0.9748982 1.0227887 1.087701
#>   21  1.087369 0.9701847 1.0278792 1.085413
#>   22  1.091994 0.9701597 1.0279274 1.085307
#>   23  1.093619 0.9701396 1.0279720 1.085214
#>   24  1.095532 0.9701195 1.0280091 1.085136
#>   25  1.095532 0.9701195 1.0280091 1.085136
#>   26  1.105450 0.9839290 1.0270110 1.078459
#>   27  1.103627 0.9870229 1.0286064 1.093806
#>   28  1.106693 0.9900265 1.0308957 1.092766
#>   29  1.107630 0.9900816 1.0349881 1.093092
#>   30  1.108245 0.9873588 1.0379898 1.092734
#>   31  1.109908 0.9873939 1.0376484 1.092713
#>   32  1.111456 0.9873187 1.0375965 1.091760
#>   33  1.113657 0.9873488 1.0375965 1.091305
#>   34  1.113691 0.9873739 1.0381641 1.091273
#>   35  1.113210 0.9906483 1.0382829 1.091108
#>   36  1.113541 0.9905379 1.0389285 1.091069
#>   37  1.113679 0.9904728 1.0396408 1.091058
#>   38  1.113524 0.9904677 1.0396408 1.091040
#>   39  1.113446 0.9904677 1.0396408 1.091020
#>   40  1.113446 0.9904627 1.0396446 1.091000
#>   41  1.113459 0.9904778 1.0396446 1.090749
#>   42  1.113541 0.9904728 1.0396446 1.090732
#>   43  1.113700 0.9904728 1.0396446 1.090714
#>   44  1.113937 0.9904677 1.0396446 1.090696
#>   45  1.114250 0.9904627 1.0396446 1.090676
#>   46  1.114268 0.9904778 1.0396446 1.090662
#>   47  1.114663 0.9904728 1.0396446 1.090646
#>   48  1.115136 0.9904728 1.0396446 1.090628
#>   49  1.115686 0.9904677 1.0396446 1.090609
#>   50  1.115699 0.9904828 1.0396446 1.090835
#>   51  1.116331 0.9904778 1.0396446 1.090819
#>   52  1.117040 0.9904728 1.0396446 1.090802
#>   53  1.117823 0.9904677 1.0396446 1.090784
#>   54  1.117836 0.9904828 1.0396446 1.091250
#>   55  1.118704 0.9904778 1.0396446 1.091234
#>   56  1.119641 0.9904728 1.0396446 1.091217
#>   57  1.120660 0.9904677 1.0396446 1.091199
#>   58  1.120677 0.9904828 1.0396483 1.091910
#>   59  1.121774 0.9904778 1.0396446 1.091893
#>   60  1.122943 0.9904728 1.0396446 1.091876
#>   61  1.124190 0.9904677 1.0396446 1.091857
#>   62  1.124211 0.9904778 1.0396483 1.092814
#>   63  1.124233 0.9904427 1.0404423 1.092789
#>   64  1.125548 0.9904377 1.0404423 1.092780
#>   65  1.126932 0.9904326 1.0404423 1.092770
#>   66  1.128394 0.9904276 1.0404423 1.092760
#>   67  1.129929 0.9904226 1.0404423 1.092748
#>   68  1.131537 0.9904176 1.0404423 1.092737
#>   69  1.133218 0.9904126 1.0404423 1.092723
#>   70  1.133218 0.9904126 1.0404423 1.092723
#>   71  1.136283 0.9941081 1.0415479 1.095697
#>   72  1.136463 0.9940630 1.0450579 1.091699
#>   73  1.135389 0.9923481 1.0455068 1.083929
#>   74  1.134868 0.9923581 1.0469167 1.078022
#>   75  1.133948 0.9922829 1.0469093 1.072208
#>   76  1.135642 0.9923331 1.0484602 1.069843
#>   77  1.134378 0.9939026 1.0494360 1.063712
#>   78  1.133420 0.9923030 1.0499147 1.058052
#>   79  1.133630 0.9941132 1.0510537 1.052377
#>   80  1.133807 0.9957228 1.0525564 1.046945
#>   81  1.134017 0.9941332 1.0535916 1.046939
#>   82  1.134013 0.9928746 1.0553874 1.043923
#>   83  1.133230 0.9914355 1.0568010 1.046730
#>   84  1.132495 0.9899312 1.0571646 1.046754
#>   85  1.131769 0.9899362 1.0582740 1.041453
#>   86  1.131064 0.9884670 1.0582740 1.036257
#>   87  1.130393 0.9884570 1.0593908 1.031200
#>   88  1.130419 0.9884419 1.0605261 1.026247
#>   89  1.130449 0.9869978 1.0605373 1.021628
#>   90  1.130440 0.9870179 1.0620622 1.020998
#>   91  1.130458 0.9870028 1.0632717 1.016612
#>   92  1.130492 0.9856038 1.0632606 1.011972
#>   93  1.130496 0.9856038 1.0644590 1.011975
#>   94  1.130496 0.9856038 1.0656797 1.011975
#>   95  1.130501 0.9842650 1.0656760 1.011978
#>   96  1.130505 0.9842650 1.0669153 1.011980
#>   97  1.130501 0.9842650 1.0681693 1.011980
#>   98  1.130509 0.9829713 1.0681656 1.011983
#>   99  1.130509 0.9829713 1.0694383 1.011984
#>   100 1.130509 0.9829713 1.0694383 1.011984
#> 
#> $median
#>      par
#> cycle       Ka        Ke         V       lag
#>   1   1.000000 1.0000000 1.0000000 1.0000000
#>   2   1.038949 1.0007399 1.0001711 1.0373792
#>   3   1.056356 1.0011429 1.0006366 1.0451696
#>   4   1.041015 1.0011414 1.0029613 0.8019811
#>   5   1.041121 1.0011414 1.0026957 0.8009597
#>   6   1.041121 1.0011414 1.0026957 0.8009597
#>   7   1.059780 1.0012739 1.0028076 0.8008284
#>   8   1.059452 1.0011317 1.0029763 0.8223454
#>   9   1.059406 1.0010848 1.0030481 0.8245007
#>   10  1.071045 1.0010414 1.0030829 0.8245001
#>   11  1.071045 1.0010414 1.0030829 0.8245001
#>   12  1.070467 1.0010935 1.0010607 0.8237037
#>   13  1.095161 1.0010905 1.0010695 0.8237280
#>   14  1.095201 1.0010908 1.0011125 0.8237467
#>   15  1.095216 1.0010908 1.0011180 0.8237754
#>   16  1.095232 1.0010908 1.0011235 0.8238043
#>   17  1.095300 1.0010908 1.0011468 0.8239262
#>   18  1.095300 1.0010908 1.0011468 0.8239262
#>   19  1.095385 0.9983918 1.0011909 1.0311191
#>   20  1.089188 0.9977229 1.0011961 1.0318178
#>   21  1.083216 0.9975292 1.0011691 1.0317251
#>   22  1.120238 0.9975762 1.0011677 1.0317215
#>   23  1.058279 0.9976165 1.0011676 1.0317215
#>   24  1.058238 0.9976512 1.0011676 1.0317215
#>   25  1.058238 0.9976512 1.0011676 1.0317215
#>   26  1.057940 0.9975933 1.0010401 1.0316305
#>   27  1.057229 0.9976430 1.0013704 1.0321766
#>   28  1.057269 0.9977312 1.0011796 1.0318285
#>   29  1.057652 0.9976696 1.0010877 1.0315678
#>   30  1.057773 0.9976087 1.0010960 1.0027777
#>   31  1.045462 0.9976088 1.0010949 1.0025997
#>   32  1.033242 0.9976088 1.0010940 1.0022664
#>   33  1.033244 0.9976088 1.0011047 0.9735329
#>   34  1.033288 0.9976088 1.0010383 0.9733523
#>   35  1.032997 0.9976088 1.0014041 0.9745905
#>   36  1.033156 0.9976088 1.0012218 0.9738684
#>   37  1.033235 0.9976088 1.0011118 0.9735174
#>   38  1.032927 0.9976088 1.0011101 0.9734962
#>   39  1.032932 0.9976088 1.0011085 0.9734745
#>   40  1.032936 0.9976088 1.0011069 0.9734524
#>   41  1.032911 0.9976088 1.0011183 0.9146421
#>   42  1.032917 0.9976088 1.0011163 0.9146200
#>   43  1.032922 0.9976088 1.0011143 0.9145974
#>   44  1.032928 0.9976088 1.0011124 0.9145743
#>   45  1.032933 0.9976088 1.0011105 0.9145507
#>   46  1.032909 0.9976088 1.0011214 0.8851918
#>   47  1.032915 0.9976088 1.0011191 0.8851681
#>   48  1.032921 0.9976088 1.0011168 0.8851440
#>   49  1.032927 0.9976088 1.0011146 0.8851193
#>   50  1.032903 0.9976088 1.0011252 0.8557531
#>   51  1.032910 0.9976088 1.0011226 0.8557284
#>   52  1.032917 0.9976088 1.0011200 0.8557031
#>   53  1.032924 0.9976088 1.0011174 0.8556773
#>   54  1.032900 0.9976088 1.0011276 0.8263046
#>   55  1.032908 0.9976088 1.0011246 0.8262787
#>   56  1.032915 0.9976088 1.0011218 0.8262522
#>   57  1.032923 0.9976088 1.0011189 0.8262252
#>   58  1.032901 0.9976088 1.0011284 0.7962166
#>   59  1.032909 0.9976088 1.0011252 0.7962075
#>   60  1.032917 0.9976088 1.0011220 0.7961978
#>   61  1.032925 0.9976088 1.0011189 0.7961876
#>   62  1.032905 0.9976088 1.0011277 0.7962155
#>   63  1.032998 0.9976088 1.0010564 0.7962303
#>   64  1.033007 0.9976088 1.0010549 0.7962230
#>   65  1.033016 0.9976088 1.0010533 0.7962153
#>   66  1.033025 0.9976088 1.0010518 0.7962073
#>   67  1.033034 0.9976088 1.0010503 0.7961990
#>   68  1.033042 0.9976088 1.0010488 0.7961904
#>   69  1.033050 0.9976088 1.0010473 0.7961816
#>   70  1.033050 0.9976088 1.0010473 0.7961816
#>   71  1.020805 0.9976186 0.9891048 0.7962937
#>   72  1.020872 0.9975908 0.9890490 0.7962379
#>   73  1.020927 0.9975859 0.9889958 0.8256311
#>   74  1.020923 0.9975922 0.9889964 0.8256339
#>   75  1.020929 0.9975986 0.9889971 0.8256358
#>   76  1.020963 0.9975640 0.9889983 0.8256386
#>   77  1.020930 0.9975722 0.9889988 0.8256454
#>   78  1.020930 0.9975805 0.9889993 0.8256479
#>   79  1.020812 0.9975805 0.9891373 0.8256234
#>   80  1.020869 0.9975805 0.9890815 0.8256238
#>   81  1.020912 0.9975805 0.9890392 0.8256284
#>   82  1.020942 0.9975805 0.9890028 0.8550647
#>   83  1.020939 0.9975805 0.9788675 0.8550346
#>   84  1.020931 0.9975805 0.9788722 0.8550464
#>   85  1.020929 0.9975805 0.9788727 0.8550483
#>   86  1.020928 0.9975805 0.9788727 0.8550561
#>   87  1.020933 0.9975805 0.9788703 0.8550893
#>   88  1.020942 0.9975805 0.9788661 0.8551291
#>   89  1.020952 0.9975805 0.9787411 0.8551763
#>   90  1.020941 0.9975805 0.9727789 0.8551529
#>   91  1.020947 0.9975805 0.9712060 0.8551850
#>   92  1.020958 0.9975805 0.9711862 0.8552348
#>   93  1.020957 0.9975805 0.9711862 0.8552348
#>   94  1.020957 0.9975805 0.9711862 0.8552348
#>   95  1.020957 0.9975805 0.9711862 0.8552348
#>   96  1.020957 0.9975805 0.9711862 0.8552348
#>   97  1.020957 0.9975805 0.9711862 0.8552348
#>   98  1.020957 0.9975805 0.9711862 0.8552348
#>   99  1.020956 0.9975805 0.9711862 0.8552348
#>   100 1.020956 0.9975805 0.9711862 0.8552348
#> 
#> $aic
#>        1        2        3        4        5        6        7        8 
#> 517.7858 491.7396 470.9806 467.1167 466.9258 466.9258 465.0927 463.5185 
#>        9       10       11       12       13       14       15       16 
#> 463.1660 463.0964 463.0964 459.0480 458.9315 458.8429 458.7916 458.7521 
#>       17       18       19       20       21       22       23       24 
#> 458.7225 458.7225 457.1578 456.3396 456.0331 456.0099 455.9982 455.9937 
#>       25       26       27       28       29       30       31       32 
#> 455.9937 454.8695 454.4178 453.8688 453.4884 453.2816 453.1872 453.1603 
#>       33       34       35       36       37       38       39       40 
#> 453.1359 453.1163 452.9602 452.7117 452.5954 452.5926 452.5911 452.5907 
#>       41       42       43       44       45       46       47       48 
#> 452.5880 452.5833 452.5799 452.5778 452.5770 452.5725 452.5675 452.5639 
#>       49       50       51       52       53       54       55       56 
#> 452.5617 452.5588 452.5525 452.5477 452.5443 452.5413 452.5340 452.5282 
#>       57       58       59       60       61       62       63       64 
#> 452.5239 452.5192 452.5112 452.5047 452.4998 452.4958 452.4864 452.4789 
#>       65       66       67       68       69       70       71       72 
#> 452.4729 452.4684 452.4652 452.4635 452.4631 452.4631 452.1157 451.9126 
#>       73       74       75       76       77       78       79       80 
#> 451.8277 451.7759 451.6906 451.6191 451.5580 451.4955 451.4427 451.3352 
#>       81       82       83       84       85       86       87       88 
#> 451.1979 451.1314 450.9686 450.8731 450.8533 450.8467 450.8133 450.8010 
#>       89       90       91       92       93       94       95       96 
#> 450.7896 450.7577 450.7407 450.7203 450.6981 450.6951 450.6755 450.6563 
#>       97       98       99      100 
#> 450.6560 450.6336 450.6168 450.6168 
#> 
#> $bic
#>        1        2        3        4        5        6        7        8 
#> 532.0070 505.9609 485.2018 481.3380 481.1471 481.1471 479.3139 477.7397 
#>        9       10       11       12       13       14       15       16 
#> 477.3873 477.3177 477.3177 473.2692 473.1527 473.0641 473.0129 472.9733 
#>       17       18       19       20       21       22       23       24 
#> 472.9437 472.9437 471.3791 470.5608 470.2543 470.2312 470.2194 470.2150 
#>       25       26       27       28       29       30       31       32 
#> 470.2150 469.0907 468.6390 468.0901 467.7097 467.5029 467.4084 467.3816 
#>       33       34       35       36       37       38       39       40 
#> 467.3571 467.3375 467.1814 466.9329 466.8166 466.8138 466.8123 466.8119 
#>       41       42       43       44       45       46       47       48 
#> 466.8092 466.8046 466.8012 466.7990 466.7982 466.7938 466.7888 466.7852 
#>       49       50       51       52       53       54       55       56 
#> 466.7829 466.7801 466.7738 466.7690 466.7656 466.7625 466.7552 466.7494 
#>       57       58       59       60       61       62       63       64 
#> 466.7452 466.7405 466.7324 466.7259 466.7210 466.7171 466.7077 466.7002 
#>       65       66       67       68       69       70       71       72 
#> 466.6941 466.6896 466.6865 466.6847 466.6844 466.6844 466.3369 466.1339 
#>       73       74       75       76       77       78       79       80 
#> 466.0489 465.9972 465.9119 465.8404 465.7792 465.7168 465.6639 465.5564 
#>       81       82       83       84       85       86       87       88 
#> 465.4192 465.3526 465.1899 465.0944 465.0746 465.0679 465.0345 465.0223 
#>       89       90       91       92       93       94       95       96 
#> 465.0108 464.9789 464.9619 464.9416 464.9193 464.9163 464.8967 464.8776 
#>       97       98       99      100 
#> 464.8772 464.8549 464.8381 464.8381 
#> 
#> attr(,"class")
#> [1] "PMcycle" "list"   
names(cycle)
#> [1] "names"  "cycnum" "ll"     "gamlam" "mean"   "sd"     "median" "aic"   
#> [9] "bic"