Summarize NPAG or IT2B Final Cycle Population Values

Extracts final cycle information from NPAG or IT2B run.

Usage

makeFinal(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeFinal is a list of class PMfinal, which contains the following:

popPoints

(NPAG only) Dataframe of the final cycle joint population density of grid points with column names equal to the name of each random parameter plus prob for the associated probability of that point

popMean

The final cycle mean for each random parameter distribution

popSD

The final cycle standard deviation for each random parameter distribution

popCV

The final cycle coefficient of variation (SD/Mean) for each random parameter distribution

popVar

The final cycle variance for each random parameter distribution

popCov

The final cycle random parameter covariance matrix

popCor

The final cycle random parameter correlation matrix

popMedian

The final cycle median values for each random parameter

#'

postPoints

(NPAG only) Dataframe of posterior population points for each of the first 100 subject, with columns id, point, parameters and probability. The first column is the subject, the second column has the population point number, followed by the values for the parameters in that point and the probability.

postMean

A nsub x npar data frame containing the means of the posterior distributions for each parameter.

postSD

A nsub x npar data frame containing the SDs of the posterior distributions for each parameter.

postVar

A nsub x npar data frame containing the variances of the posterior distributions for each parameter.

postCov

NPAG only: An array of dimensions npar x npar x nsub that contains the covariances of the posterior distributions for each parameter and subject.

postCor

NPAG only: An array of dimensions npar x npar x nsub that contains the correlations of the posterior distributions for each parameter and subject.

postMed

A nsub x npar data frame containing the medians of the posterior distributions for each parameter.

shrinkage

A data frame with the shrinkage for each parameter. popVar is comprised of variance(EBE) + variance(EBD), where EBE is the Emprical Bayes Estimate or mean of the posterior distribution for the parameter. EBD is the Empirical Bayes Distribution, or the full Bayesian posterior distribution. In other words, if Bayesian posterior distributions are wide for a given parameter due to sparse or uninformative sampling, then most of the population variance is due to this variance and shrinkage of the EBE variance is high because individual posterior estimates shrink towards the population mean.

gridpts

(NPAG only) Initial number of support points

nsub

Number of subjects

ab

Matrix of boundaries for random parameter values

A plot method exists in plot for PMfinal objects.

Details

This function will parse the output of NPparse or ITparse to generate a list suitable for analysis and plotting of NPAG or IT2B final cycle population values.

See also

Author

Michael Neely

Examples

data(NPdata.1)
#> Warning: data set ‘NPdata.1’ not found
final <- makeFinal(NPdata.1)
#> Error in makeFinal(NPdata.1): object 'NPdata.1' not found
final
#> Error in eval(expr, envir, enclos): object 'final' not found
names(final)
#> Error in eval(expr, envir, enclos): object 'final' not found
plot(final)
#> Error in plot(final): object 'final' not found
data(ITdata.2)
#> Warning: data set ‘ITdata.2’ not found
final2 <- makeFinal(ITdata.2)
#> Error in makeFinal(ITdata.2): object 'ITdata.2' not found
final2
#> Error in eval(expr, envir, enclos): object 'final2' not found
names(final2)
#> Error in eval(expr, envir, enclos): object 'final2' not found
plot(final2)
#> Error in plot(final2): object 'final2' not found