Generated observed vs. predicted data

Generates an observed vs. predicted data.frame from an NPAG or IT2B object

Usage

makeOP(data)

Arguments

data

A suitable data object of the NPAG or IT2B class (see NPparse or ITparse).

Value

The output of makeOP is a data frame of class PMop, which has a population and posterior prediction object (also class PMop) for each output equation. Each of these has 13 columns:

id

subject identification

time

observation time in relative hours

obs

observation

pred

prediction

pred.type

Population predictions based on Bayesian prior parameter value distribution, or individual predictions based on Bayesian posterior parameter value distributions

icen

Predictions based on mean or median of Bayesian pred.type parameter values

outeq

output equation number

block

dosing block number for each subject, as defined by dose resets (evid=4).

obsSD

standard deviation of the observation based on the assay error polynomial

d

prediction error, pred-obs

ds

squared prediction error

wd

weighted prediction error, which is the prediction error divided by the obsSD

wds

weighted squared prediction error

A plot method exists in plot for PMop objects.

Details

makeOP will parse the output of NPparse or ITparse to generate a data.frame suitable for analysis and plotting of observed vs. population or or individual predicted outputs.

See also

Author

Michael Neely

Examples

data(NPdata.1)
#> Warning: data set ‘NPdata.1’ not found
op <- makeOP(NPdata.1)
#> Error in makeOP(NPdata.1): object 'NPdata.1' not found
op
#> Error in eval(expr, envir, enclos): object 'op' not found
names(op)
#> Error in eval(expr, envir, enclos): object 'op' not found
summary(op)
#> Error in summary(op): object 'op' not found
plot(op)
#> Error in plot(op): object 'op' not found