Plot Pmetrics Final Cycle Parameter Value Distributions

Plot R6 PM_final objects made by makeFinal and loaded as a field in the PM_result object, e.g. PM_result$final.

Usage

# S3 method for PM_final
plot(
  x,
  formula,
  marker = T,
  line = T,
  density = F,
  grid = T,
  xlab,
  ylab,
  standardize
)

Arguments

x

The name of an PM_final data object as a field in a PM_result R6 object, e.g PM_result$final.

formula

An optional formula of the form y ~ x, where y and x are two model parameters to plot in a 3-dimensional bivariate plot. See details.

marker

See details for which objects the marker argument controls. This argument maps to the plotly marker object. It can be boolean or a list. TRUE will plot the marker with default characteristics. FALSE will suppress marker plotting. If a list, can control many marker characteristics, including overriding defaults. Use the plotly plotly::schema() command in the console and navigate to traces > scatter > attributes > marker to see all the ways the marker can be formatted. Most common will be:

color Fill color for NPAG bars, marker color for bivariate NPAG plots. Ignored for IT2B plots.

symbol Plotting character. See plotly::schema(), traces > scatter > attributes > marker > symbol > values. Only relevant for bivariate NPAG plots.

size Character size in points. Only relevant for bivariate NPAG plots.

opacity Bar fill color for univariate NPAG plots or marker opacity for bivariate NPAG plots. Ignored for IT2B plots. Ranges between 0 (fully transparent) to 1 (fully opaque).

line A list of additional attributes governing the outline for bars in univariate NPAG plots or markers in bivariate NPAG plots. Ignored for IT2B plots.

color Outline color. Default is "black".

width Outline width. Default is 1. Example: marker = list(color = "red", symbol = "triangle", opacity = 0.8, line = list(color = "black", width = 2))

line

Controls characteristics of lines. This argument maps to plotly trace line attributes. TRUE will plot default lines. FALSE will suppress lines. If a list, can control many line characteristics, including overriding defaults. Use the plotly plotly::schema() command in the console and navigate to traces > scatter > attributes > line to see all the ways the line can be formatted. Most common will be:

color Line color.

dash Plotting character. See plotly::schema(), traces > scatter > attributes > line > dash > values.

width Thickness in points. Example: line = list(color = "red", dash = "longdash", width = 2)

density

Boolean operator to plot a kernel density function overlying the histogram of a univariate marginal parameter distribution from NPAG; the default is FALSE. See density. Ignored for IT2B output.

grid

Controls grid display. This argument maps to the xaxis and yaxis layout objects in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis or yaxis > gridcolor or gridwidth. It is a Boolean operator. If FALSE, no grid is plotted. If TRUE, the default color grey50 and width 1 will be plotted at major tick marks. If a list, color and width can be customized. Examples:

grid = F

grid = list(gridcolor = "black", gridwidth = 2)

xlab

Define x-axis label for bivariate NPAG or IT2B plot. Default is the name of the plotted x-variable.

ylab

Define y-axis label for bivariate NPAG or IT2B plot. Default is the name of the plotted y-variable.

standardize

Standardize the normal parameter distribution plots from IT2B to the same scale x-axis. Ignored for NPAG output.

Value

Plots the object.

Details

If formula is omitted, this will generate a marginal plot for each parameter. For NPAG data, this will be a histogram of marginal values for each parameter and the associated probability of that value. For IT2B, this will be a series of normal distributions with mean and standard deviation equal to the mean and standard deviation of each parameter marginal distribution. IF formula IS specified, this will generate one of two plots.

On the other hand, if formula is two parameters, e.g. CL~V, this will generate a bivariate plot. For NPAG data, it will be support point with size proportional to the probability of each point. For IT2B, it will be an elliptical distribution of a bivariate normal distribution centered at the mean of each plotted variable and surrounding quantiles of the bivariate distribution plotted in decreasing shades of grey.

The line argument is used to format:

the density line drawn from an NPAG PM_final object if density = T

the drop lines from an NPAG PM_final object when a formula is specified to generate a bivariate plot

the lines drawing the normal distribution of parameter values from an IT2B PM_Final object. The marker argument is used to format:

the bar fill color and opacity, and the outline color and width from an NPAG PM_final object

the markers used from an NPAG PM_final object when a formula is specified to generate a bivariate plot

See also

Author

Michael Neely

Examples

#NPAG
NPex$final$plot()

NPex$final$plot(density = T)

NPex$final$plot(Ke ~ V)

#IT2B
ITex$final$plot()

ITex$final$plot(Ke ~ V)
#> Error in x$popMean[1, c(whichX, whichY)]: incorrect number of dimensions