Plot PM_pta Percent Target Attainment objects

Plots PM_pta objects

Usage

# S3 method for PM_pta
plot(
  x,
  include,
  exclude,
  type = "pta",
  mult = 1,
  log = F,
  outeq = 1,
  line = T,
  marker = T,
  grid = T,
  legend = T,
  ci = 0.9,
  xlab,
  ylab,
  xlim,
  ylim,
  ...
)

Arguments

x

The name of an *PM_pta" data object read by makePTA

include

A vector of simulations (regimens) to include in the plot, e.g. include = c(1,3)

exclude

A vector of simulations (regimens) in the plot, e.g. exclude = c(2,4:6)

type

Character vector controlling type of plot. Default is "pta", which plots proportion with success on the y-axis and target on the x-axis. The other choice is "pdi", which plots the median pdi (pharmacodynamic index), e.g. AUC/MIC, on the y-axis, and target on the x-axis.

mult

Multiplication factor for y axis, e.g. to convert mg/L to ng/mL

log

Boolean operator to plot the y axis in log base 10. This argument maps to the the yaxis type value in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > yaxis > type. Example: log = T

outeq

Which output equation to plot. Default is 1.

line

Controls characteristics of lines. This argument maps to the plotly line object. It can be boolean or a list. TRUE will plot the line with default characteristics for each simulated regimen. FALSE will suppress line plotting. If a list, it functions a little differently than other Pmetrics plotly functions. Rather than controlling individual line characteristics, for this plot, the line argument should be a list of the options for group based plotting, where each group corresponds to a simulated regimen. The possible elements of the line list should be exactly named:

color Maps to the plot_ly colors argument to override default colors applied to the lines for each regimen. This can be a named palette, which can be obtained with RColorBrewer::display.brewer.all() or a vector of hexadecimal color names. One way to ensure reliable color palettes is to use the ColorBrewer site. Choosing the number of data classes to correspond to regimens, and qualitative data results in a distinct palette. Easiest importing into R is to copy/paste the Export of JavaScript on the ColorBrewer website. The default is "Spectral". Palettes with fewer colors than regimens will be recycled. A color can also be a character vector of color names, recycled as needed. For example, a print-friendly choice is line = list(color = "black").

width Maps to the plot_ly width argument to override default widths applied to the lines for each regimen. All lines will have the same width. The default value is 2.

dash Maps to the plot_ly linetypes argument to override default styles applied to the lines for each regimen. If numeric, will map to lty par values. It can also be a character vector of dash names as listed in plot_ly. Example: line = list(color = "Blues", width = 1, dash = 2), whicb will result in dotted lines (dash = 2) all with width 1 but in different shades of blue.

marker

Controls the plotting symbol. This argument maps to the plotly marker object. It can be boolean or a list. TRUE will plot the profiles with default characteristics for each simulated regimen. FALSE will suppress line plotting. If a list, it functions a little differently than other Pmetrics plotly functions. Rather than controlling individual marker characteristics, for this plot, the marker argument should be a list of the options for group based plotting, where each group corresponds to a simulated regimen. The possible elements of the marker list should be exactly named:

color Default marker color is the same as the line color. If line color is specified, marker color does not need to also be specified. Even if line plotting is suppressed with line = F, the default color value of "Spectral" will be applied to markers, unless specified, e.g. marker = list(color = "Blues").

symbol Maps to the plot_ly symbols argument to override default symbols applied to the markers for each regimen. If only one value is supplied for this, it will be recycled for each regimen, i.e. all will have the same symbol. See plotly::schema(), traces > scatter > attributes > marker > symbol > values for options.

size Maps to the plot_ly size argument to override default size applied to the markers for each regimen. All markers will have the same size. The default value is 12.

grid

Controls grid display. This argument maps to the xaxis and yaxis layout objects in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis or yaxis > gridcolor or gridwidth. It is a Boolean operator. If FALSE, no grid is plotted. If TRUE, the default color grey50 and width 1 will be plotted at major tick marks. If a list, color and width can be customized. Examples:

grid = F

grid = list(gridcolor = "black", gridwidth = 2)

legend

Controls display of legend. This argument maps to the plotly layout showlegend and legend arguments. It is either a boolean operator (most common) or a list of parameters to be supplied to plotly. See plotly::schema() > layout > layoutAttributes > legend and showlegend for more details on the available options for formatting. If legend is supplied as a list, the plotly layout > layoutAttributes > showlegend value will be set to TRUE automatically. Examples:

legend = T

legend = list(orientation = "h", font = list(color = "blue"))

ci

Confidence interval around curves on type = "pdi" plot, on scale of 0 to 1. Default is 0.9.

xlab

Label for the x axis. Default is "Target" when targets are discrete, and "Regimen" when targets are sampled.

ylab

Label for the y axis. Default is "Proportion with success" for plot type = "pta" and "Pharmacodynamic Index" for plot type = "pdi".

xlim

Limits of the x axis as a vector. This argument maps to the the xaxis range in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis > range. Example: xlim = c(0,1)

ylim

Limits of the y axis as a vector. This argument maps to the the yaxis range in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > yaxis > range. Example: ylim = c(0,100)

...

Other parameters which can be passed to control axis formatting. Specify as lists named either xaxis or yaxis. See plotly::schema() layout > layoutAttributes > xaxis/yaxis for options. Example: xaxis = list(title = list(text = "Time (h)", font = list(family = "Arial", size = 14, color = "black")))

Value

Plots the object.

Details

This function will plot the percent target attainment for objects made with the makePTA function. For the legend, defaults that are different that the standard are:

x Default "topright"

legend Default will be the labeled regimen names supplied during makePTA, or if missing, "Regimen 1, Regimen 2,...Regimen n", where n is the number of regimens in the PM_pta object. This default can be overridden by a supplied character vector of regimen names.

col The color of each Regimen plot as specified by the default color scheme or col

pch The plotting character for each Regimen plot as specified by the default plotting characters or pch

lty The line type of each Regimen plot as specified by the default line types or lty

bg Default "white"

See also

Author

Michael Neely