Plot Pmetrics Simulation Objects

Plots PM_sim objects with the option to perform a visual and numerical predictive check

Usage

# S3 method for PM_sim
plot(
  x,
  mult = 1,
  log = T,
  probs = list(0.05, 0.25, 0.5, 0.75, 0.95),
  ci = 0.95,
  binSize = 0,
  outeq = 1,
  marker = F,
  grid = F,
  xlab = "Time",
  ylab = "Output",
  xlim,
  ylim,
  legend = F,
  obs,
  quiet = F,
  ...
)

Arguments

x

The name of an PM_sim data object generated by PM_sim

mult

Multiplication factor for y axis, e.g. to convert mg/L to ng/mL

log

Boolean operator to plot the y axis in log base 10. This argument maps to the the yaxis type value in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > yaxis > type. Example: log = T

probs

List of quantiles to plot, e.g. probs = list(0.05, 0.5, 0.95). If set to NA, all simulated profiles will be plotted, and numerical predictive checking will be suppressed. To format individual quantile lines, expand the list object for that quantile For example, probs = list(0.05, list(value = 0.5, color = "red", width = 2, dash = "dash"), 0.95). Use the plotly plotly::schema() command in the console and navigate to traces > scatter > attributes > line to see all the ways the marker can be formatted. Most common will be:

color Fill color and outline color if not otherwise specified

symbol Plotting character. See plotly::schema(), traces > scatter > attributes > marker > symbol > values.

size Character size in points

opacity Ranging between 0 (fully transparent) to 1 (fully opaque).

ci

Width of confidence interval bands around simulated quantiles, from 0 to 1. If 0, or nsim<100, will not plot. Default is 0.95, i.e. 95th percentile with tails of 2.5 percent above and below excluded.

binSize

Width of binning interval for simulated concentrations, in time units, e.g. hours. A binSize of 0.5 will pull all simulated concentrations +/- 0.5 hours into the same time. This is useful for plotting PMsim objects made during makeValid. The default is 0, i.e. no binning. If an obs object is provided, it will be binned similarly.

outeq

Which output equation to plot. Default is 1.

marker

Controls the plotting symbol for observations. This argument maps to the plotly marker object. It can be boolean or a list. TRUE will plot the marker with default characteristics. FALSE will suppress marker plotting. If a list, can control many marker characteristics, including overriding defaults. Use the plotly plotly::schema() command in the console and navigate to traces > scatter > attributes > marker to see all the ways the marker can be formatted. Most common will be:

color Marker color.

symbol Plotting character. See plotly::schema(), traces > scatter > attributes > marker > symbol > values.

size Character size in points.

opacity Ranging between 0 (fully transparent) to 1 (fully opaque).

line A list of additional attributes governing the outline for filled shapes, most commonly color and width. Example: marker = list(color = "red", symbol = "triangle", opacity = 0.8, line = list(color = "black", width = 2))

grid

Controls grid display. This argument maps to the xaxis and yaxis layout objects in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis or yaxis > gridcolor or gridwidth. It is a Boolean operator. If FALSE, no grid is plotted. If TRUE, the default color grey50 and width 1 will be plotted at major tick marks. If a list, color and width can be customized. Examples:

grid = F

grid = list(gridcolor = "black", gridwidth = 2)

xlab

Value for x axis label. This argument maps to the the xaxis layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis to see other ways to customize this axis using lists as additional arguments to the plot function. Example: xaxis = list(title = list(text = "Time", font = list(color = "red", family = "Arial", size = 10)))

ylab

Value for y axis label. This argument maps to the yaxis layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > yaxis to see other ways to customize this axis using lists as additional arguments to the plot function. Example: yaxis = list(title = list(text = "Concentration", font = list(color = "blue", family = "Times New Roman", size = 12)))

xlim

Limits of the x axis as a vector. This argument maps to the the xaxis range in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > xaxis > range. Example: xlim = c(0,1)

ylim

Limits of the y axis as a vector. This argument maps to the the yaxis range in the layout object in plotly. Use the plotly plotly::schema() command in the console and navigate to layout > layoutAttributes > yaxis > range. Example: ylim = c(0,100)

legend

Controls display of legend. This argument maps to the plotly layout showlegend and legend arguments. It is either a boolean operator (most common) or a list of parameters to be supplied to plotly. See plotly::schema() > layout > layoutAttributes > legend and showlegend for more details on the available options for formatting. If legend is supplied as a list, the plotly layout > layoutAttributes > showlegend value will be set to TRUE automatically. Examples:

legend = T

legend = list(orientation = "h", font = list(color = "blue"))

obs

The name of a PM_result data object or the PM_op field in the PM_result object, all generated by PM_load. For example, if run1 <- PM_load(1) and sim1 is a PM_sim object, then sim1$plot(obs = run1) or sim1$plot(obs = run1$op). If specified, the observations will be overlaid upon the simulation plot enabling a visual predictive check. In this case, a list object will be returned with two items: $npc containing the quantiles and probability that the observations are below each quantile (binomial test); and $simsum, the times of each observation and the value of the simulated quantile with upper and lower confidence intervals at that time. Additionally, the number of observations beyond the 5th and 95th percentiles will be reported and the binomial test P-value if this number is different than the expected 10% value.

quiet

Suppress plot if TRUE.

...

Other parameters which can be passed to control axis formatting. Specify as lists named either xaxis or yaxis. See plotly::schema() layout > xaxis/yaxis for options.

Value

Plots the simulation object. If obs is included, a list will be returned with the folowing items:

npc A dataframe with three columns: quantile, prop_less, pval. ** quantile are those specified by the probs argument to the plot call ** prop_less are the proportion of simulated observations at all times less than the quantile ** pval is the P-value of the difference in the prop.less and quantile by the beta-binomial test.

simsum A dataframe with the quantile concentration at each simulated time, with lower and upper confidence intervals

obs A data frame similar to an PMop object made by makeOP at the end of a run with the addition of the quantile for each observation

Details

Simulated observations are plotted as quantiles on the y-axis vs. time on the x.axis. If measured observations are included, a visual and numerical predictive check will be performed. The default plot is to omit markers, but if the marker arugment is set to TRUE, the resulting marker will have the following plotly properties: list(symbol = "circle-open", color = "black", size = 8). By default a grid is omitted. The legend is also omitted by default, but if included, clicking on a quantile item in the legend will hide it in the plot, and double clicking will hide all other quantiles.

See also

Author

Michael Neely