Plot Pmetrics Validation Objects

Plot PMvalid objects

Usage

# S3 method for PMvalid
plot(
  x,
  type = "vpc",
  tad = F,
  icen = "median",
  outeq = 1,
  lower = 0.025,
  upper = 0.975,
  log = F,
  pch.obs = 1,
  col.obs = "black",
  cex.obs = 1,
  data_theme = "color",
  plot_theme = theme_grey(),
  col.obs.ci = "blue",
  col.obs.med = "red",
  col.sim.ci = "dodgerblue",
  col.sim.med = "lightpink",
  axis.x = NULL,
  axis.y = NULL
)

Arguments

x

The name of an PMvalid data object generated by makeValid.

type

Default is “vpc” for a visual prective check, but could be “pcvpc” for a prediction-corrected visual predictive check.

tad

Plot using time after dose if TRUE. Default is FALSE which plots using standard relative time. This will be the only option if tad was not set to TRUE when making the PMvalid object.

icen

Can be either “median” for the predictions based on medians of the population parameter value distributions, or “mean”. Default is “median”.

outeq

The number of the output equation to simulate/test. Default is 1.

lower

The lower quantile displayed for the observed and simulated profiles. Default is 0.025.

upper

The upper quantile displayed for the observed and simulated profiles. Default is 0.975.

log

Boolean operator to plot in semilog space. The default is FALSE.

pch.obs

Control the plotting character used for observations. Default is 1, i.e. an open circle. See points for other values of pch.

col.obs

Color for observations. Default is black.

cex.obs

Size for observatins. Default is 1.

col.obs.ci

Color of the observation confidence interval (set by lower and upper). Default is blue.

col.obs.med

Color of the observation median. Default is red.

col.sim.ci

Color of the simulation confidence interval (set by lower and upper). Default is dodgerblue.

col.sim.med

Color of the simulation median. Default is lightpink.

theme

Default is “color”, but could be “grey” or “gray”.

xlab

Label for x axis. Default is “Time”.

ylab

Label for y axis. Default is “Observation”.

xlim

Limits for x axis

ylim

Limits for y axis

...

Not currently used

Value

Plots the object using ggplot2.

See also

Author

Michael Neely