Extract CDC pediatric growth charts

Will extract height and weight for boys, girls or both for a given range of ages in months and percentile. This can be useful for simulations in Pmetrics.

Usage

qgrowth(
  sex = c("M", "F", "B"),
  percentile = c("5", "10", "25", "50", "75", "90", "95"),
  agemos = (seq(0, 18) * 12)
)

Arguments

sex

A single quoted character: “M” for males, “F” for females, or “B” for both, in which case an average of the two sexes will be returned. Default is “M”.

percentile

An integer of the percentile for each age/sex to return. Default is 5.

agemos

The ages in months to return. The default is seq(0,18)*12, i.e. 1 to 18 years.

Value

A dataframe with columns

age

Age in months

wt

Weight in kilograms

ht

Height or length in centimeters

sex

The selected sex

percentile

The selected percentile

Author

Michael Neely