Summary Statistics for PMfinal Objects

Generates summary statistics of final population model parameters.

Usage

# S3 method for PMfinal
summary(object, ..., lower = 0.025, upper = 0.975)

Arguments

object

The PMfinal object made after an NPAG or IT2B, e.g. final.1 after run 1.

...

Other parameters which are not necessary.

lower

Desired lower confidence interval boundary. Default is 0.025. Ignored for IT2B objects.

upper

Desired upper confidence interval boundary. Default is 0.975. Ignored for IT2B objects.

Value

The output is a data frame. For NPAG this has 4 columns:

value

The value of the summary statistic

par

The name of the parameter

type

Either WtMed for weighted median, or MAWD for MAWD (see details)

quantile

Requested lower, 0.5 (median), and upper quantiles

For IT2B this has 5 columns:

mean

Parameter mean value

se.mean

Standard error of the mean

cv.mean

Error of the mean divided by mean

var

Variance of the parameter values

se.var

Standard error of the variance

Details

For NPAG runs, this function will generate weighted medians as central tendencies of the population points with a 95\ and the median absolute weighted deviation (MAWD) from the median as a measure of the variance, with its 95\ 95\ sample from a normal distribution. To estimate these non-parametric summaries, the function uses a Monte Carlo simulation approach, creating 1000 x npoint samples with replacement from the weighted marginal distribution of each parameter, where npoint is the number of support points in the model. As an example, if there are 100 support points, npoint = 100, and for Ka, there will be 1000 sets of 100 samples drawn from the weighted marginal distribution of the values for Ka. For each of the 1,000 sets of npoint values, the median and MAWD are calculated, with MAWD equal to the median absolute difference between each point and the median of that set. The output is npoint estimates of the weighted median and npoint estimates of the MAWD for each parameter, from which the median, 2.5th, and 97.5th percentiles can be found as point estimates and 95\ interval limits, respectively, of both the weighted median and MAWD.

For IT2B runs, the function will return the mean and variance of each parameter, and the standard errors of these terms, using SE (mean) = SD/sqrt(nsub) and SE (var) = var * sqrt(2/(nsub-1)).

See also

Author

Michael Neely

Examples

data(NPdata.1)
#> Warning: data set ‘NPdata.1’ not found
final <- makeFinal(NPdata.1)
#> Error in makeFinal(NPdata.1): object 'NPdata.1' not found
summary(final)
#> Error in summary(final): object 'final' not found
data(ITdata.2)
#> Warning: data set ‘ITdata.2’ not found
final2 <- makeFinal(ITdata.2)
#> Error in makeFinal(ITdata.2): object 'ITdata.2' not found
summary(final2)
#> Error in summary(final2): object 'final2' not found